samples: nrf_desktop: hw_interface: Fix TWI in buttons handler
After a switch to new TWI driver communication with SX1509 started
to fail. It seems that when multiple write transactions are scheduled
in one call, stop bit is included even if requested not to.
To mitigate the problem we switch to TWI API for reading and
writing device registers. This is actually the correct way to do
this communication in the first place.

Jira:DESK-239

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>