samples: nrf_desktop: disable NCS subdirectory include
NCS subdirectory include was added to avoid depending on temp hack
added to Nordic's fork of Zephyr, that adds NCS directory to Zephyr
build tree.
Unfortunately zephyr libraries can be build only if construction is
defined within the Zephyr build tree. If zephyr libraries are
defined outside of it (e.g. together with application) they are
not picked up and fail to link.

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>