samples: nrf_desktop: motion interface for pca10056
Now that pca10056 is one of the targets for the nRF Desktop application,
proper HW interfaces can be defined for this board. The first interface
is used to initialize the board at the start and uninitialize it before
going to system off. The second one is used to emulate mouse motion with
DK buttons.

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>