subsys: Bluetooth: services: Align HID to multicontext
This contains HID service adapted to change context
for all connections and an aligned example of HID Mouse
for changes related to addind multiple contexts to HID
service. The use of an example with two connections
has been shown. Also, the example of NRF Destop
has been aligned with the new HID API.

Signed-off-by: Kamil Gawor <Kamil.Gawor@nordicsemi.no>