samples: nrf_desktop: Allow to send zero delta motions
Motion sensor is collecting data before it is read. It could happen
that when moved rapidly both ways the sum of motions will be zero.
We could avoid sending this report, switch sensor data fetching off
and wait for irq to report new movement. The problem is this takes
time and breaks the cursor motion smoothness. Notice that user is
constantly moving the mouse in such scenario.

The new approach is to allow one readout with zero values to be sent
by the sensor hw interface. When no motion is detected for the second
time interface is switched to irq (idle) mode. It seems to work
smoothly.

Jira:DESK-280

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>