cmake: nrf_desktop: Modernize how CONF_FILE is set
The nrf_desktop application needs to set CONF_FILE differently
depending on the board. But the board is not easily knowable at the
time when CONF_FILE needs to be set.

To work around this chicken-and-egg problem we defer evaluation with a
macro hook. This is the recommended and standard way of maintaining a
complex board-dependent conf file.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>