bluetooth: hci: Fix hci_driver defines in controller only build
In a controller only build, BT_RX_PRIO and BT_RX_STACK_SIZE is not
defined. If BT_RX_PRIO is defined we need to use that priority to avoid
deadlock in the host, if not we are free to choose a priority.
BT_RX_STACK_SIZE is meant for the host RX thread, so use our own
definition instead

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>