drivers: net: Added new driver to transfer Ethernet over RTT
During development and testing we need reliable connection for
network traffic. Radio based solutions depends on other modules that
are complex, so more things may go wrong. Existing SLIP driver uses
serial port, so logs must be disabled. RTT is simple and does not
require any peripherals, so it is best for this job.

Signed-off-by: Dominik Kilian <Dominik.Kilian@nordicsemi.no>