cmake: git: Make nrfxlib a required dependency of nrf
nrfxlib has been an optional dependency of fw-nrfconnect-nrf until
now, this patch makes it a required dependency instead.

When NCS is used it will consist of a set of mandatory and optional
dependencies. Until now fw-nrfconnect-zephyr and fw-nrfconnect-nrf
have been mandatory dependencies, while nrfxlib has been optional.

When NCS is distributed it will be simplest for users and contributors
of NCS if nrfxlib is always distributed with it. For instance, having
nrfxlib always present means that NCS users that enable their first
nrfxlib component will not have to first download nrfxlib, but instead
be able to directly enable it through the Kconfig configuration system
as if it was an integrated part of NCS (which it is intended to be).

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>