subsys: profiler: declare log_event_buf contents conditionally
In profiler.h, the struct log_event_buf relies on
PROFILER_CUSTOM_EVENT_BUF_LEN being defined, which is not the
case if the profiler is disabled. This patch lets the contents
of log_event_buf be compiled out in such case, to prevent
errors when compiling a C file including profiler.h.

This is possible since event_manager does not uses the contents
of the log_event_buf struct directly but only passes that to the
profiler; when the profiler is disabled, the struct can be empty.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>