samples: Rename nrf_cloud sample to a more descriptive name.
This commit moves the nrf_cloud sample to a folder named nrf_cloud and
changes the name of the sample to gps_tracker. Additional nrf_cloud
based samples are in development, and it will be simpler to browse the
repository with the samples grouped together with descriptive names.

This also improves the situation where everything related to nrf cloud
is named nrf cloud. Today the online solution, the nrf91 stack, and the
sample on the nrf91 are all named nrf cloud. Giving the sample a
distinct name will improve this situation.

Signed-off-by: Rune Holmgren <Rune.Holmgren@nordicsemi.no>