samples: nrf9160: at_client: Adapt to AT host running in thread
The AT host library now processes AT commands in a dedicated
thread, and it's therefore not needed to call library functions
from main.
This patch removes inclusion of now deleted at_host.h and
the function calls to the AT host library. The main stack
is also increased to avoid stack overflow.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>