[temphack] lib: at_host: Workaround for spinning socket polling thread
This patch adds a workaround that puts the socket polling function
to sleep for a short while whenever poll() times out. Currently,
poll() times out immediately, regardless of what the timeout
is set to, causing this thread to spin without yielding to
lower priority threads, such as the logging thread.

This fixes the issue where logging is not working when AT host
is enabled at the same time.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>