lib: bsdlib: prevent compilation if firmware is not non-secure
bsdlib requires to be run from non-secure firmware.
Emit a very clear error message and abort compilation if a project
that does not define CONFIG_TRUSTED_EXECUTION_NONSECURE tries to
compile bsdlib glue files.

The rationale for this change is to make it easy to prevent
the situation where a project that uses bsdlib is compiled for
the wrong board, e.g. nrf9160_pca10090 instead of the non-secure
variant nrf9160_pca10090ns. The project would compile just fine
and run into a hard-fault at runtime, or emit a cryptic or
seemingly unrelated error during compilation.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>