build: Switch to using Zephyr modules to include the repo
In order to include the repository in the Zephyr build, we used to rely
on a custom hook in the zephyr repository's root CMakeLists.txt and
Kconfig.zephyr files. Now that Zephyr natively supports external
modules, use this mechanism to include nrf in the Zephyr build.
This implies that we no longer need to include nrfxlib from the nrf
repository, since that is done automatically by Zephyr's build system
through a separate commit in nrfxlib.

Includes the corresponding update to west.yml in order to switch to the
correct zephyr and nrfxlib revisions.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>