subsys: bluetooth: scan: add possibility to scan manufacturer data
This makes it possible to scan for chuncks of data in the
manufacturer data advertisement packages. It can be used to, e.g.,
scan for iBeacons. The chuncks of data can be less then the actual
advertisement package, e.g., if somebody just wants to scan for the
iBeacon's UUID, but not for major, minor etc.

Signed-off-by: Christian Hirsch <christian.hirsch@tuwien.ac.at>