subsys: spm: add a SystemReset secure service
This commit implements a simple Secure Service that
allows the Non-Secure firmware to issue a system reset.
This is done by exposing a secure entry function, which
can be called by the Non-Secure firmware.

Add implementation of sys_arch_reboot() that calls into the SPM
when in non-secure firmware.

Add a Kconfig for secure services and configure SPM to
expose secure services by default.

This commit modifies the SPM core, so that it does not
allow anymore NVIC SystemReset to be issued from Non-Secure
environment.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>
Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>