subsys: bootloader: fix dependency for key generation and signing
The dependencies of generating and signing the firmware image were
incorrect. This meant that every time you did a rebuild, it would sign
the image again and create a mismatch between the provisioned data and
the firmware image. This commit fixes it, and the build will no
longer try to generate new certificates when doing a rebuild with clean
files.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>