subsys: bootloader: Fix for making b0 build for nrf51
Since we don't have a proper build step for building the uart
driver in nrf51 we need to disable debug when building for nrf51. So
this commit moves the configuration from the sample into the default
configurations of Kconfig so debug in bootloader will be enabled for
everything except nRF51.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>