samples: nrf9160: mqtt_simple: Add print of IPv4 address using ntop
Printing out the resolved IPv4 address isn't really helpful unless you
manually convert it afterwards. This change introduces the use of utils
from zephyr's `net` subsystem and uses the `ntop` function to convert
it from binary to an `\0` terminated string which is easier to read when
it is printed out.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>