manifest: update zephyr
The reason for having a pull/#/head in master was that there was a
dependency from the fw-nrfconnect-zephyr repo to the fw-nrfconnect-nrf
repo. Hence the fw-nrfconnect-nrf PR had to be merged first, so that the
CI for the fw-nrfconnect-zephyr PR could be run. Now that both PRs are
in, and CI is green for both, lets change to the latest SHA of
fw-nrfconnect-zephyr.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>