spm: Add a random number secure service
Based on MbedTLS.
Add initialization function for secure_services.
Refactor calculation of secure/non-secure split in SPM because if
MCUboot is present, SPM might start in the middle of a region, and
span two regions.

Also refactor the reboot code to remove unnecessary docs and function.
Add file-level doc string.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>