subsys/bluetooth/gatt_pool: Reimplement of GATT pool
This commit removes the usage of CONFIG_BT_GATT_HIDS_ATTR_MAX
from the gatt_pool internal code.

The GATT pool was totally reimplemented that it has its own object.
Functions was renamed to match mempool interface more.

Also the attribute freeing was automated and simplified as
the user does not have to know what type of argument was allocated.

JIRA: NCSDK-2123

Signed-off-by: Radoslaw Koppel <radoslaw.koppel@nordicsemi.no>
Signed-off-by: Bartosz Gentkowski <bartosz.gentkowski@nordicsemi.no>