net: lib: download_client: reconnect on connection closed
The maximum number of requests that can be served through one
keep-alive connection usually defaults to 100, after which
the connection is closed by the server.

This change re-introduces the capability of the library to
detect when the server includes the field: "Connection: close"
in the response header, and attempts to reconnect automatically.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>