net: lib: http_download: rename from download_client
Rename `download_client` to `http_download` and move it under /net.

This library downloads files over HTTP and HTTPS, so let's give it
a name that better conveys what it does; since it's a network
library move the header under /net from the include root folder.

In the documentation, we mention this as the `HTTP downloader`
library, which has a better ring to it than `HTTP download`.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>