lib: bsdlib: fix wrong and deprecated UARTE macros
The DT_NORDIC_NRF_UARTE_N_* family of macros is deprecated.

Plus, they refer to the N-th instance of a compatible DTS
node, the order of which is solely dependent on the Python
script that generates the definitions. Thus, the macros
may not refer to the N-th UARTE hardware instance.

Use instead the correct family of macros for the N-th
UARTE hardware instance: DT_NORDIC_NRF_UARTE_UART_N.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>