samples: nrf9160: lwm2m_client: sample demonstrating LwM2M
This sample demonstrates the use of Zephyr's LwM2M client configured
for the nRF9160 DevKit to support: Buttons, Switches, Buzzer,
Connectivity Data and more.

To support FOTA capability the FIRMWARE object is implemented to
allow CoAP transmission of a signed combined binary of the secure
and non-secure samples.   MCUboot checks the signature on the next
boot and then moves that binary from slot1 to slot0 where it will
be booted.

Signed-off-by: Michael Scott <mike@foundries.io>