applications: asset_tracker: Add real GPS control
This patch adds a GPS controller module and adapts the
application to use it. The controller is intended to handle
LTE link configuration and synchronize it with the GPS
driver.

Environment data kan only be sent when the GPS is inactive,
but this window may not always overlap with the environment
send timer. Therefore, send environment data when GPS fix
is acquired to ensure environment data is sent at all.
Also introduce a back-off time for retrying to send until
the GPS becomes inactive.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>
Signed-off-by: Bernt Johan Damslora <bernt.johan.damslora@nordicsemi.no>