applications: nrf_desktop: usb_state: Clear report pending flag
Will fix problem with lack of notification about report sent in case
of sudden disconnection of USB. In such situation driver may be
disconnected before report is sent. That lead usb_state to be in an
unfinished report sent state and caused rise of assertion when new
report request arrived.

Jira:DESK-651

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>