cmake: add offset to secondary slot to pm target
This is needed in MCUBoots CMakeLists. The reason it is done
like this is that these values are not available in
configure time. Hence, the offset would be calculated by
doing arithmatics on genexp values, which would end up with
an expression like (0xa + 0xb) which is not accepted as
an argument by objcopy.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>