drivers: nrf9160_gps: Add option to handle AT commands manually
Adds the NRF9160_GPS_HANDLE_MODEM_CONFIGURATION option.  This provides
the ability to remove the at_cmd library dependency from the GPS driver
in case someone's application is managing the AT socket in a custom
driver.  This obviously comes with the caveat that with this in the
non-default setting the application must separately send the proper AT
commands to configure the GPS.

Signed-off-by: Justin Brzozoski <justin.brzozoski@signal-fire.com>