bluetooth: gatt_pool: Adapt to upstream CCC storage changes
Adapt to upstream changes to the way CCC storage is handled. In order
to control the storage of the CCC configuration data we need use the
BT_GATT_CCC_MANAGED macro. This allows us to use the CONTAINER_OF
to get the surrounding structure of the CCC configuration data.

NCSDK-2977

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>