aws_fota_json: fix bug where null-terminator was not copied from json
strlen does not account for the null-terminator, but this must also
be copied to the output buffers, as there is no length data being
exposed.

Verify this by updating the tests:
Prior to this, the test used strncmp, and was not explicitly
initialized. Both these issues could end up in bugs in the code
related to null termination not being caught.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>