lib: at_notif: Added module.
A new module, at_notif, has been added as a library. The module
dispatches events to registered modules, which are identified by
a context pointer and a callback function.

Modules can register or deregister subscriptions of AT-command
notifications through public functions.

The module is scheduled to initialize after at_cmd.

Signed-off-by: PK Chan <pak.kee.chan@nordicsemi.no>