dfu: support parsing paths for B1 MCUBoot upgrades
When updating the second stage bootloader, B1, (MCUBoot in this case),
two variants of the new firmware is required. One for each possible
flash placement (slot), S0 and S1. To present the two variants,
the path will contain two entries, separated by a space. The entry
before the space contains the path to the S0 candidate, while the
entry after the space contains the path to the S1 candidate.

This commit adds a function which parses a space separated path,
and given a flag indicating which candidate should be used (S0 or S1)
it returns a pointer to the correct part of the path.

In addition, this commit updates fota_download.c to find the active
slot, and indicate to the path parsing function which variant it
should point to.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>