lib: modem_info: Fix area code string termination
This patch fixes an issue where the area code string was null
terminated too early. The area code is 'a two-byte Tracking Area
Code (TAC) in hexadecimal format', which is represented in a
four-byte ASCII string. Null-termination is therefore moved from
byte 2 to 4. Also a cast to double is removed.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>