bootloader: bl_validation: Fix looping through public keys
break out of loop on -EHASHINV instead of -ESIGINV. -EHASHINV is
returned if the provided validation info is using the wrong public
key (which is what we want). This is because the provision area stores
the public key hash, so the public key can be validated even before
checking the signature.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>