samples: nrf9160: aws_fota: Fix device shadow topic out of memory
The device shadow topic was cut short due to the size calculations being
of by 1. Since the length is different than what it was in the AWS Fota
library. This resulted in the device shadow topic returning -12(ENOMEM).

Also fixes a compiler warning when compiling with `USE_NRF_CLOUD` where
it would say `update_device_shadow declared but not used`.

Fixes NCSDK-3551

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>