Bluetooth: gatt_dm: Fix assert in bt_gatt_discover on start handle
Fix assert in bt_gatt_discover API on start_handle and end_handle.
The GATT Discovery manager would provide the invalid attribute handle 0
for the start handle at the end of the discovering all services.
At this point the end_handle is 0xffff, so when start handle is set to
end handle + 1 it would cause an integer overflow to 0.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>