applications: asset_tracker: Increase default MQTT keep-alive
Increasing deafult MQTT keep-alive time to 1200 seconds,
because short keep-alive times may be too short when using power
saving modes, such as PSM and eDRX, in combination with
low-bandwidth NB-IoT networks.

In addition, the poll() timeout is set to 1/3 of the MQTT
keep-alive value. This will ensure that we check the need for
sending a ping and maintaining connection to cloud solution in
time for setting up slow connections and send the ping in time.
The underlying mqtt_live() checks time since last activity,
and only sends ping if it's been a period equal to the keep-
alive value since last data was sent. This means we risked to
send ping close 2 * CONFIG_MQTT_KEEPALIVE, and even more
delayed on NB-IoT. With this patch, we  will not go past
4/3 * CONFIG_MQTT_KEEPALIVE, which is usually accpeted by
the cloud service.

Fixes NCSDK-3608

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>