drivers: lte_lc: Ignore non-CEREG notifications when connnecting
After the at_notif module was added, the link controller has
parsed all notifications intended for other users than the
link controller as if they were CEREG notifications, and
printing errors when it turns out they're not.
This patch fixes the issue by silently ignoring every notification
that's not a CEREG response.
It also removes logging all incoming notifications, as they are
often not relevant, for the same reasons.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>