doc: Improve PM docs
Various improvements to the PM docs.

Don't state that PM will be disabled when the build strategy is set to
'skip building', as this is incorrect.

Use the term child image instead of sub-image, as this is the correct
terminology.

Don't mention that the .config file is a Kconfig fragment as this is
an internal technical detail that is subject to change.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>