lib: modem_info: Make date/time readout optional
The `AT+CCLK?` command issued by modem_info module to retireve real-time
clock value from the modem can fail in certain circumstances (for
instnace if network time is not available in the modem yet), causing the
entire modem_info readout to fail.

This commit makes this parameter optional, so that this readout can be
disabled if the clock value is not used.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>