scripts: remove NCS internals from requirements.txt
The editdistance and pygit2 dependencies are only needed by
maintainers for the internal ncs-loot and ncs-compare west extension
commands.

They're also causing installation problems for Windows users, as e.g.
editdistance is a C extension which is not currently available in
wheel form on Windows + Python 3.8.

Let's move them out to their own requirements file to make this
problem go away. This has the side benefit of making the installation
footprint for regular users smaller.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>