partition_manager: allow alignment of first partition
Alignment of partitions is done in two stages, before and after the
'app' partition. All partitions before 'app' is aligned by being 'moved
up' in flash by inserting an empty partition before them.

This is however problematic for the first partition, as this contains
the vector table, and needs to be stored at address 0.

To solve this issue, include some special handling when aligning the
end of the first partition so that the empty partition is inserted
after it instead of before.

Note that as opposed to regular partitions behind the 'app' partition,
the first partition is not moved down when inserting an empty partition
behind it.

This commit introduces the functionality mentioned above.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>