net: lib: fota_download: Fix when last fragment write fails
When writing the last fragment fails the `dfu_target_done`is still
called applying the update. This is because two events are sent, both
`DOWNLOAD_CLIENT_EVT_FRAGMENT` and `DOWNLOAD_CLIENT_EVT_DONE` even if we
return a non-zero value. This is because `dl->progress == dl->file_size`
still applies.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>