samples: bluetooth: throughput: Fix scan active during throughput test
Fix throughput test not achieving the expected throughput.
The occurred because the scanner was active during the throughput test
as bt_le_create_conn had already been called when the slave connection
was established.

It was not possible to cancel this connection object in the connecting
state because the same object is used to reference the slave connection.

This happens because the test is trying to establish connections in both
master and slave role at the same time. This use case is not supported
by the zephyr host, and can lead to connection object confusion as two
contexts are connected to the same bluetooth device. The API
bt_conn_lookup_addr_le does not work for this scenario.

Also fixed issues where wrong variable was used for err, mixing u8_t and
int, as well other cases where err was checked without being assigned,
or should have been checked.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>