bluetooth: hci_driver: Update to new signature of bt_read_static_addr
Update bt_read_static_addr signature in nrfxlib hci_driver to use the
vendor specific HCI command struct. This signature had already been
changed and is not part of Zephyrs API. Which caused the wrong identity
address to be used, and the random address type bits was set
incorrectly.

The new signature includes reading the Identity Root which allows us to
have a Bluetooth identity without storing it in flash.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>