scripts: make west ncs-compare use NCS manifest-revs
This means that even if the NCS revision in west.yml is not a
SHA (e.g. if it's pull/*/head), this command will still work as long
as you've run 'west update'

The benefit is that the command can always work robustly. The downside
is that you need to run 'west update' to compare.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>