samples: nrf9160: Add nRF Cloud A-GPS sample
Adds a minimal sample showing hos nRF Cloud and A-GPS can be
used together.
Fetches A-GPS data upon connection to cloud, and starts GPS
when the data is stored to the modem. Searches for fix for
a maximum of 2 minutes, and restarts search every 4 minutes.
Button press sends a short message to cloud. Long press, 3 secs,
reboots the device.
Sending {"reboot":true} from the cloud to the device also reboots.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>