lib: at_cmd: Added putting NULL byte in response string
There is already a check to make sure the buffer has enough room for
the entire string plus one -- a strict greater-than check -- so there
isn't a reason this can't terminate the string itself. Since it doesn't
send back the length of the response, it puts a heavy burden on users
of the AT command module to have to completely clear out their buffers
each time they send another command.

Signed-off-by: Erik Johnson <erik.johnson@nimbelink.com>