lib: bsdlib: Add Kconfig option to allow running BSDLib in secure
This allows a user for BSDLib to utilize the full chip and all it's
peripherals. The whole RAM is available except the block which is
occupied by statically allocating the size of memory needed and putting
it at the start of the location BSDLib expects it to be through the
`__attribute__` hint and linker script sections.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>