fw_info: Add fw_info_ext_api_provide() as an EXT_API.
When B0 boots an image, it first checks whether to give it access to
some functions in it own code. This is what is called an EXT_API. When
it finds a request from the child image, it tries to fulfill it by
seeing what APIs are provided by itself, or - maybe in the future -
other images. It does this with a function called
fw_info_ext_api_provide(). The problem arises when e.g. MCUboot is the
child image, and itself boots a child image. B0 is not able to fulfill
EXT_API requests from this "grand-child" image because a) it's not aware
that it exists, and b) EXT_APIs are shared through RAM, so it has to
happen immediately before booting the image.

This patch provides the ext_api_provide() function as an EXT_API so that
MCUboot can do B0's job of fulfilling the app's EXT_API requests before
booting it.

Also, it makes some small improvements to fw_info_ext_api_provide():
- Handle NULL fw_info to allow the return value of fw_info_find() to be
  directly used.
- Check that the found EXT_API doesn't come from the same image as
  requested it.
- Print warning when optional EXT_API_REQs fail.

and Kconfig.template.fw_info_ext_api:
- Remove explicit "default n" so default can be overridden later.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>