doc: add a note about asset_tracker boot chain
I personally lost some time doing some development related to this
application due to an incorrect assumption that it was using the
secure bootloader chain which is documented in ug_bootloader.rst.

In fact, by default, this application does not use the secure
bootloader chain.

In case others are confused by this, try to help by adding a short
paragraph explaining this.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
Signed-off-by: Ruth Fuchss <ruth.fuchss@nordicsemi.no>