partition_manager: Add more descriptive error message on non compliance
If the C-preprocessor is not able to find a match for a symbol it will
leave the symbol name. This caused partition_manager to fail which it
should. However the error it raised was not very descriptive so I added
a function which checks the alignment symbols and if they are not
integers a `TypeError` exception will be raised with an additional
print out on what failed.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>
Co-Authored-By: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>