cmake: pattern match '_ns' and 'ns' to non-secure boards
Prior to this a hard coded list was used to identify non-secure boards.

With this patch it is possible to define OOT boards that are non-secure
by having the _ns or ns suffix.

Note that it is not supported to have a board name ending with
'_ns' or 'ns' if it is not a trustzone enabled board.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>