sample: http_application_update: Add erase in progress and done events
HTTP application update did not use the `ERASE_IN_PROGRESS` event and
`ERASE_DONE`. This caused the sample to either hang on modem firmware
delete or use a long time. This change makes the sample shutdown the LTE
link on the `ERASE_IN_PROGRESS` event so that the modem has more time to
process the erase request. When `ERASE_DONE` is received the modem
reconnects the LTE link.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>