applications: connectivity_bridge: Add "UART always on" option
Add "UART always on" option that will keep UART RX always enabled,
even if there is no listener (USB or BLE).
This is useful to catch log messages that happened before the COM port
or BLE Service was opened/connected.

Note the following consequences when this option is enabled:
- UART RX will always draw current
- If USB CDC buffer fills up, new data will overflow,
  leaving only the oldest data in the buffer.

Also Updated the config.txt code to make it easier to add
new boolean options.

Signed-off-by: Audun Korneliussen <audun.korneliussen@nordicsemi.no>