samples: central_bas: handle directed advertising
The central_bas sample is now able to handle directed advertising.
Application connects to every device which uses it.

In Readme, users are recommended to test this sample with
peripheral_hids_mouse, so this change makes both applications more
compatible with each other.

This commit is based on the similar change done for the central_hids
sample: 4c87bfc83f61792c681ed50c5affabfb22266233.

Ref: NCSDK-4647

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>