net: lib: aws_fota: add persistent sessions support
Add to aws_jobs.c/.h functions which only create, not also subscribe to,
various topics.

In aws_fota.c, fix logging of job_doc, and log it even in good case. Log
the fota_update_topic. Detect session present flag in evt handler and
save for later. Skip subscribing to the fota topics if session is valid,
and only create the names, and advance to next operation that would
otherwise occur on a SUBACK.

Jira: TG91-248

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>