drivers: entropy_cc310: patch in with new devicetree API
In upstream Zephyr commit 43a7d26603711cd60f516f9b731615ef1f82f6d1,
CONFIG_ENTROPY_NAME was removed. This Kconfig option was replaced by
adding a new "zephyr,entropy" property of the devicetree /chosen node.

The rest of the system now uses the label property of the chosen
zephyr,entropy node to find the proper device structure to use for
system entropy, including:

- low level kernel initialization
- some bluetooth link layer and controller code
- some network socket TLS handling code

We need to continue to interoperate with those, so we can't keep using
an ENTROPY_NAME Kconfig option in NCS.

However, we don't always have a devicetree node we can point to
to instantiate this entropy device:

- in secure configurations in an SoC with TrustZone or SoCs without
  TZ, we have a real 'cryptocell' node, and can use its label

- in non-secure configurations for SoCs with TZ, however, this file
  delegates Zephyr's entropy driver API callbacks to the NCS platform
  library for accessing the CC310 via SPM: there is no hardware
  node within the devicetree to use for the struct device name

To resolve this situation for now, use the cryptocell node's label
property if it exists, but fall back on a "cryptocell_sw" node (that's
made available in NCS) in the other case.

We may need a better long-term solution for non-secure code.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>