Bluetooth: hci_driver: Fix deadlock when system runs out of evt bufs
Fix deadlock between the hci_core rx_thread and the hci_driver
rx_thread.

The hci_core rx_thread is during processing of the encryption change
event will send new HCI commands to the controller and waits on a
semaphore for the command complete event.
The hci_core tx_thread successfully sends this to the controller so
it is now up to the hci_driver rx_thread to fetch the command complete
event.
The hci_driver rx_thread is waiting for an available buffer for HCI
events after having used up all the HCI event buffers from the pool
and is now waiting for the hci_core rx_thread to free up buffers.

Handle this by marking advertising reports and the QoS conn event
reports as discardable events. In a combined build this make the
hci_driver retrieve the buffers for these events for a different HCI
event buffer pool. Also do not wait for available buffers for these
events but discard if no buffer is available.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>