lib: lte_lc: Do not return error if AT+CSCON fails
AT+CSCON was introduced in nRF9160 modem firmware version v1.1.0.
Earlier versions will return ERROR when the command is received,
which would cause w_lte_lc_init() return with an error code.
Failed AT+CSCON will result in RRC mode change notifications
not being received. The only effect for the link controller users
is that they will not receive RRC events, which they would anyway not
expect due to the too-old modem firmware for this functionality.

If application FOTA is performed on a device with a modem
older than v1.1.0, calls to lte_lc_init() and
lte_lc_init_and_connect() would fail without this patch,
resulting in a device without LTE connection and likely no way to
recover the error with FOTA.

Because of this, failed AT+CSCON should not return an error code,
and this patch removes that behavior.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>