net: lib: fota_download: Add option to define host port number
The download client already supports connecting to a port number other
than the default HTTP/HTTPS port. However, this functionality was not
exposed in `fota_download_start`. This adds the functionality of
defining the host port number when trying to download a file with
`fota_download` in case you are not using the typical 80 or 443 ports.

Signed-off-by: Tiago Costa <tiagosmcosta@gmail.com>