doc: Updated the doc_build.rst with Zephyr CMake package changes
This commit adjust the guideline for building documentation.
With the introduction of Zephyr CMake package it is no long needed to
source zephyr-env.sh or execute zephyr-env.cmd.

The documentation is also update to request cmake to create the build
folder, which remove the need for creating this folder manually.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>