partition_manager: add MBR offset for nrf52840dongle
This is needed to correctly mirror functionality seen before
including a subsys enabled the partition manager.

Specifically, the nrf52840dongle will either use an
nRF5 SDK based bootloader or MCUBoot to program
applications. If the nRF5 SDK based bootloader
is used, we need to make room for the MBR at
the beginning of flash.

Ref: NCSDK-5551
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>