hal_nordic: Fix undefined reference to z_bt_ctlr_used_nrf_ppi_channels
Update hal_nordic revision to fix a problem in nrfx_glue that causes
an undefined reference error when CONFIG_BT_CTLR is enabled but
CONFIG_BT_LL_SW_SPLIT is not.
Remove also a no longer needed workaround for this problem that was
used in the connectivity_brigde application.

NCSDK-5589

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>