net: lib: nrf_cloud: Add option to use non-blocking socket
Using a blocking socket can in case of LTE network loss cause
the user of blocking socket operations hang indefinitely, or
until a watchdog timer kicks in.
This patch adds the option to configure the nRF Cloud socket
as non-blocking to work around this issue, at the cost of
sometimes giving up on sending data that could have been
sent with a short period of blocking.

Fixes NCSDK-5642

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>