doc: switch to Zephyr max-width style
Zephyr overrides the "read the docs" standard size of the content pane,
which is 800px, to use all of the available space. For the nRF
documentation generation this property was not overridden resulting in
a fixed size pane. This applies the same change that Zephyr does for
all documentation.

JIRA: NCSDK-1018

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>