doc: add css rule to hide struct/union #includes
breathe automatically adds a "#include ..." line for every
class/struct/union, and this behavior cannot be disabled at the moment.
The added lines are missing the path, which makes them less useful. The
issue happens because doxygen does not add the file path to the
particular XML path breathe consumes to generate the page information.
An alternative solution is to just hide those "#include" lines through a
CSS property, which is being done here.

JIRA: NCSDK-4850

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>