cmake: Expose PM_DOMAINS through GLOBAL PROPERTY
This is done to avoid problems when re-configuring projects.

Previously, the variable was exposed as a CACHE variable
to make it available for other cmake files.

When exposed through the CACHE the value will not change
when the project is reconfigured.

This is resolved by exposing the variable through a GLOBAL
PROPERTY, which is not retained between builds.

Ref: NCSDK-5474

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>