doc: tools: created minimum required and recommended tools
This commit add requirement text files for tools such as
- Python
- CMake
- GNU ARM Embedded
- SES
and more

Those files can be shared between CI and the Toolchain manager repo to
ensure those are using same versions of tools when building.

A page in doxygen has been created to present the main version of
those tools to users of NCS.

All Python packages required for building applications/samples and
documentations are described in independent tables.

Ref: NCSDK-4914

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>