bluetooth: services: adding Alexa Gadgets service sample
Alexa Gadgets is a framework for connecting wireless gadgets to
Amazon Echo, using Bluetooth LE and other protocols.
Communication with the Echo is implemented as a stream format
(packet fragmentation) through a custom GATT service.
Commands and data is encoded in a protobuf format.

This pull request implements the Alexa Gadgets GATT Service,
and a sample which implements the Alexa Gadgets Profile.
Protobuf encoding/decoding is handled at the profile level,
while the GATT Service manages the stream format.

The protobuf format is defined in an external Github repo,
as a collection of .proto files.
These proto files must be compiled using the protoc protobuf compiler
and the nanopb plugin. The nanopb package must be manually downloaded
or otherwise made available to CMake (as described in README.rst).
The CMake configuration for this sample will fetch the .proto files
and do the protoc compilation automatically.

Signed-off-by: Audun Korneliussen <audun.korneliussen@nordicsemi.no>

doc: review for Alexa Gadgets

Changes in RST for the service and sample.

Signed-off-by: Bartosz Gentkowski <bartosz.gentkowski@nordicsemi.no>