doc: fix make zephyr after clean-zephyr
When running the `zephyr` target after `clean-zephyr` the build fails
because a modification time script is called to run on the Doxygen XML
output, which does not exist. It happens because Doxygen is not run due
to a timestamp file. This adds a rule to remove the file that Zephyr
uses to determine if Doxygen needs to run on clean-*.

Since `-E remove` seems to work like `rm -f` this still works for other
clean-* rules without side-effects.

JIRA: NCSDK-3012

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>