doc: add a hide search matches option
This brings in the searchbox from sphinx, which also adds a "Hide Search
Matches" button below the search bar, to remove the search elements
highlighting.

This also adds a static js with a callback to allow rewriting the url
location bar to remove the search arguments, so one can copy/paste it
without submitting the hightlight terms.

JIRA: NCSDK-5245

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>