pygit2_helpers: add zephyr_commit_area()
This is copied from code I originally wrote for Foundries.io
which was in the ncs-tools repository here:

https://github.com/NordicPlayground/ncs-tools

I used to use this for making blog posts about what changed in Zephyr.
We want to start including significant Zephyr changes in the NCS
release notes as well, so it will hopefully be useful for splitting up
the work among maintainers during upmerges.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>