mpsl: Add missing multithreading lock
Wrap mpsl_low_priority_process with a multithreading lock to
avoid the following race condition:

1. The application calls ble_controller_flash_write(). This API obtains
   lock. The API is called from thread priority.
2. A low priority MPSL signal becomes available.
   mpsl_low_priority_process() interrupts the call to
   ble_controller_flash_write().
   This may give a race condition in MPSL.

In the future we may ensure that the flash APIs are called from a
cooperative thread. That will eliminate the need for the lock.

The flash driver now no longer obtains the lock when continuing flash
operations from the OPERATION_COMPLETE callback, because this callback
is executed from the execution context of mpsl_low_priority_process().

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>