Bluetooth: samples: Fix bt_conn_unref missing in connected callback
Fix central samples taking a reference in the scan_connecting
callback but not releasing this reference in the connected callback
when creating the connection failed.

In this a bt_conn_unref missing in connected callback would lead to
the connection not being freed.

Also restart the scanner in this case so that the sample can find
a new advertiser.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>