cmake: multi image: pass on the DTC_OVERLAY_FILE to sub images
Device tree overlay files can be used to modify a board's device tree,
including things such as the UART TX/RX/CTS/RTS lines. This must be
passed on to sub images, especially mcuboot, so that, for example,
the logging UART uses the correct hardware pins for mcuboot.

Jira: NCSDK-5972

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>