scripts: point to validation info of last image
Prior to this commit, b0 could not boot hex files which consisted
of two or more merged hex files. This because the 'size' field in
the 'fw_info' struct did not point to the validation info at the
end of the last image in b1.

This commit adds a script which updates the 'size' field of all
fw_info structs prior to signing.

To get the value of the MAGIC field, move the inclusion of
fw_info CMakeLists.txt earlier, and expose the variable through
the CACHE.

NCSDK-4800

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>