doc: jump to the same page on version switch
When clicking on an old release message banner link, it would always
jump to the last/latest release inside the "nrf" documentation set. This
fixes the navigation so it tries to stay on the same page while going to
the last/latest versions.

When switching the nRF Connect SDK release in the dropbox also try to
stay on the same page when jumping.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>