drivers: gps: nrf9160_gps: Set GPS priority when enabled
GPS can be given priority over LTE to let it search for
satellites also in situations where the radio would usually
be in use by LTE. This is useful when the LTE network doesnn't
offer functionality like eDRX or PSM where the GPS can have full
access to the radio.

Currently a workaround has to be applied to make setting priority
work. The GPS must be started for the setting to be applicable,
but setting the priority also makes the GPS stop. Therefore
it's necessary to start the GPS once more after the priority
option has been applied.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>