net: lib: aws_iot: yield READY after subscribing to configured topics
This changes the purpose of the AWS_IOT_EVT_READY event to communicate
that the client has subscribed to all configured topics.

Previously it was a duplication of the AWS_IOT_EVT_CONNECTED
event and provided no meaningful information.

An error is returned from mqtt_evt_handler() if the subscription
fails.

The documentation is updated to reflect this change and
no longer mentions a ready state of the broker which
does not exist in AWS IoT.

Co-Authored-by: Simen Røstad <Simen.Rostad@NordicSemi.no>
Signed-off-by: Markus Tacker <Markus.Tacker@NordicSemi.no>