samples: openthread: use common configuration and features sets
This commit moves the common configuration for OpenThread samples
to a single file for ease of maintenance. This file is included
automatically for every sample as an overlay via `CMakeLists.txt`.

This commit also makes use of the predefined Nordic library
features sets defined in nrfxlib `Kconfig` for some of the samples.

Signed-off-by: Eduardo Montoya <eduardo.montoya@nordicsemi.no>