drivers: entropy: cc310: Workaround for spm secure services limitation
`spm_request_random_number()` requires that the application requests
exactly 144 random bytes, no different. As it's called directly in the
cc310 entropy driver, this is an unacceptable limitation, as the
`entropy_get_entropy()` function is used in various different
libraries, not aware of this shortcoming.

This commit works around this limitation in the
`entropy_cc310_rng_get_entropy` implementation.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>