net: fota_download: Fix disconnecting before downloading from offset
Before the library did not disconnect before trying to restarting the
download from the offset. This worked previously with how partial
downloading works. Now that the `download_client` library does not do
partial downloads for HTTP we need to disconnect and connect before
trying to start the download from another offset.

NCSDK-6279

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>