Bluetooth: scan: Update scan module for extended advertising PDUs
Update the scan module for extended advertising PDUs. We need to
determine if the advertiser is connectable by looking at the
advertising properties for extended advertising PDUs.

Instead of copying over all of the additional advertising information
provide a pointer to the scan report recv information struct,
this one includes the address as well, so remove it from the device
info struct.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>