applications: serial_lte_modem: support TLS server
This commit adds support of TLS server. The TLS server is based on
zephyr native TLS socket. To build serial LTE modem with native TLS
socket support, add the following to your west build command:

      -DOVERLAY_CONFIG=native_tls_overlay.conf

If the zephyr native TLS is enabled, the credentials are mapped to
continuous security tags in modem and stored as Root CA certificate
so that the credentials can be read from modem.

The available sec_tag in modem (0 – 2147483647) are divided by 10
to store mapped credentials:
   Root CA certificate (ASCII text) at sec_tag*10 + 0
   Client/Server certificate (ASCII text) at sec_tag*10 + 1
   Client/Server private key (ASCII text) at sec_tag*10 + 2
   Pre-shared Key (PSK) at sec_tag*10 + 3
   PSK identity (ASCII text) at sec_tag*10 + 4
   Public Key (ASCII text) at sec_tag*10 + 5

The credential management command (AT%CMNG) is overridden
to map sec_tag into modem.

Currently PSK, PSK identity and Public Key are not supported.
Also, DTLS server is not supported yet.

Signed-off-by: Larry Tsai <larry.tsai@nordicsemi.no>