net: lib: nrf_cloud: make MQTT client ID prefix configurable
The `nrf-` prefix is reserved for Nordic devices:
https://api.nrfcloud.com/v1#operation/CreateDeviceCertificate:

    If you want to create a certificate for a non-Nordic device,
    any deviceId is sufficient that does not start with
    `nrf-` (we recommend using a GUID).

This change allows customers to use the `asset_tracker` application
with nRF Connect for Cloud with their own SIPs without modifying
the source.

This is especially relevant for customers that are building
development kit like products, and allows them to work out of the
box with the stock `asset_tracker` and nRF Connect for Cloud.

Signed-off-by: Markus Tacker <Markus.Tacker@NordicSemi.no>
Co-authored-by: Sigvart Hovland <Sigvart Hovland@NordicSemi.no>