drivers: gps: nrf9160_gps: Add COEX and MAGPIO handling for Thingy:91
Adds COEX0 and MAGPIO handling for Thingy:91 to simplify use
of external GPS antenna.
In addition, AT commands to set these confiurations in the modem
are now sent when gps_init() is called instead of at boot time.
This is needed for the AT commands to be sent successfully also
in cases where BSD library is not initialized at boot.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>