boards: arm: thingy91_nrf9160: Send AT cmd only if BSDlib initialized
Adds a check for CONFIG_BSD_LIBRARY_SYS_INIT to avoid
sending AT commands if BSDlib is not initialized.
If commands sould fail to be sent, a log message is printed to inform
that MAGPIO and COEX0 can not be configured, but must instead be
handled by application, drivers or libraries before LTE/GPS is used.

Fixes CIA-114

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>