net: lib: aws_fota: Change default security tag value
This patch changes the default value for the security
tag used for FOTA downloads. The well-known value of
42 may explain the meaning of life but should not be used
as the default security tag value for FOTA downloads.
It has been proven to be rather common that developers use
the security tag number 42 to store credentials related to other
connections and by changing the default value to -1, potential
future conflicts can be avoided.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>