bluetooth: gatt_dm: report failed discovery to app
If a device turns off during discovery, the call to
bt_gatt_discover() fails in bt_gatt_dm_continue.
However, the call to discovery_complete_error() is missing,
so application code is not alerted to the failure. It can
wait forever for the scan to complete, as a result.

Add the missing call to discovery_complete_error().

Jira: NCSDK-8641

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>