net: lib: download_client: Add kconfig for range requests
Add a configuration option to always use HTTP Range requests.
This option can be useful to limit the amount of incoming data
from the server by downloading only one fragment at a time.
It increases the protocol overhead but also gives time to
the application to process the fragments as they are downloaded,
instead of having to keep up to speed while downloading the whole file.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>
Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>