net: lib: download_client: only set socket timeout for UDP sockets
Let the socket timeout be configured for UDP sockets only.

With TCP sockets and HTTP, a timeout caused a re-transmission of
the GET request which may have led to recv() reading the same header
twice, mistaking one for the actual payload. This could have either
triggered an assert or silently passed wrong data to the application
in a fragment.

Since the download is running in a separate thread, it is not very
useful to have a timeout on a TCP socket, plus the socket will
anyway return an error if the server closes the connection and no
further data is expected.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>