drivers: gps: nrf9160_gps: Don't clear TCXO offset in delete mask
The delete_apgs_data boolean is used to indicate that GNSS
assistance data should be deleted. The TCXO offset (bit 7) is not
part of the A-GPS data set that can be input to the device,
so it should not be included in the bitmask of the data
to be deleted.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>