[wip] doc: generalize Kconfig reference to 'Configuration'
      TODO: rebase the zephyr patch onto Ioannis's upmerge,
            when that is ready

Do this by also generating the devicetree bindings index in the same
docset. Add a new index file that links to both the Kconfig options
and the DT bindings index.

Do the devicetree parts in the same way as was previously done for
kconfig: by calling into the zephyr specific script, passing it a
different output directory to use.

There are also various places in the tree where we need to move files
around or tweak strings to change 'Kconfig' to 'Configuration', or
reflect the change in build system target naming.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>