net: lib: azure_fota: Start FOTA download only once
In some cases, FOTA download could be started more than once if two
or more device twin messages were received in rapid succession.
The consequence is corruption of the downloaded image.

This patch moves a state check in order to stop processing incoming
device twin messages at an earlier stage if a download is ongoing.
Moving the check prevents the state to be changed from 'downloading',
which would allow another download to start.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>