net: lib: aws_iot: Add check for client id length
Add check for client ID length to make sure that the user sets
an appropriate length of the buffer hosting the client ID. This patch
also increases the default max length of the client ID buffer. This is
due to an increasing number of cases of people using long client IDs
for their MQTT connection and not understanding quite what happens
when the library fails.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>