samples: nrf9160: Increase MQTT keepalive timeout
This patch increases the MQTT keepalive for cloud vendor specific
samples to the maximum recommended value listed in each cloud vendors
online documentation.

This is to limit the amount of packets sent to the broker which in turn
creates a more energy efficient sample than if the default keepalive
of 60 seconds were to be used.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>