samples: nrf9160: lwm2m_client: Fix FW download on non-zero offset
In case FW download is interrupted for whatever reason (for instance
board reset) we end up with partially written FW update.

Since in the PUSH mode FOTA it's not possible to notify the server it
shouldn't send the firmware from scratch, simply ignore the already
stored bytes to avoid the corruption of the already written firmware.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>