samples: nrf9160: udp: Split modem configuration into init and connect
When enabling RAI the link controller checks if the current network mode
is NB-IoT (which is the only network mode that supports this feature).
This patch moves the enabling of RAI to after the modem configuration is
initialized to ensure that NB-IoT is set as the network mode before
calling lte_lc_rai_req(). Previously this API call would fail even
though NB-IoT was configured as the network mode.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>