Bluetooth: Mesh: Fix overflow on lvl_delta
When performing a delta set in the lighting models, the delta is added
to the current value, then stored in an integer that only holds 16 bits.
If the delta change overflows the 16 bit integer, it will wrap around
when stored, instead of clamping to the min/max of the storage variable.
Although clamping is done in the set-functions, it's already too late,
as the level has overflowed its variable and wrapped around.

Adds boundary clamp before storing the target values in the 16 bit
integer for gen_plvl, lightness_srv and light_temp_srv.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>