pcd: add busy wait to fix network core update
There is an issue where the application core
will block the network core from writing to
shared RAM while it is reading a variable
in a busy loop.

Fix this issue by adding a busy_wait instead.

The update of the network core takes seconds,
so using 1 second as a wait value makes sense.

Ref: NCSDK-7289
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>