lib: date_time: Add API to check if the library has obtained valid time
Convenience function that checks if the date time library has
obtained valid time. This allows the application to check if valid time
has been obtained before attempting to call other functions of the
date time library to convert/get timestamps. This patch also adds a
test for the new function.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>