actions: labeler: Add a no-changelog rule
Whenever a Pull Request is posted, we need to check if the changes it
contains warrant an addition to the changelog.
In order to better make sure that this rule is abided by, label the PR
by default with the no-changelog label. If the PR is updated with a
change to the changelog, the label will be automatically removed by the
action. If the user determines that no changes to the changelog are
required, they can remove the label themselves, creating a record of
this action.

Pull Requests with the "no-changelog" label present should not be
merged.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>