ci: split documentation build/publish workflows
Split the current documentation build and publish workflow into two
steps, a build which runs on the "pull_request" env with no repo and
secret access and a "workflow_run" triggered by the build workflow to
publish the pages. This removes usage of pull_request_target which
could be abused by badly behavior actors.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>