net: lib: nrf_cloud: Add API to send preencoded data to nrf nrf_cloud
Add API in nrf_cloud.h to send preencoded data to nRF Cloud. Prior to
this patch the only way to send data to nRF Cloud was via:

- nrf_cloud_sensor_data_stream()
- nrf_cloud_shadow_update()
- nrf_cloud_sensor_data_send()

All of these functions encodes the data before it is sent to cloud.
(Alters the payload). The new API makes it possible for the application
to send data to nRF Cloud specific endpoint types (message, state)
without the library encoding the data first.

[CIA-207]

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>