ci: update documentation publishing
This updates the documentation publishing to use a better format to
communicate information to the backend. As well as sending the archive,
a file named `monitor_$GITHUB_RUN_ID.txt` (for uniquenes) is also added
with commands and parameters. The backend currently supports
`"publish <folder> <archive.tgz>"` for uncompressing `archive.tgz`
under `folder`, and `"remove <folder>"` to remove a tree under
`folder`.

A new worflow was add to run when a PR is merged/closed, sending
information to the backend to remove the local folder.

An extra step was added to invoke `carlescufi/action-doc-url` to comment
on the PR with the link to the published documentation. A new secret,
`NCS_DOC_URL_ROOT`, must be configured with the URL root where the
documentation will live.

Publishing of the `master` branch was also enabled.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>