applications: serial_lte_modem: add CONFIG for external XTAL usage
The use of external XTAL was added in PR#3311 following spec of UARTE in PS.
Based on test, it increases power consumption when SLM is put to Idle mode.
Add CONFIG_SLM_EXTERNAL_XTAL (default 'y') to control the use of external XTAL.
Temporarily disable the usage in prj.conf so as to save power consumption.

Signed-off-by: Jun Qing Zou <jun.qing.zou@nordicsemi.no>