lib: at_cmd_parser: support neagive numbers
There are some usecases where the numbers returned from the AT
interface is negative. These negative numbers should be supported by the
AT parser. This update should be safe because there is already a bug in
the parser code that converts a uint32_t to a int and back again.

Signed-off-by: Even Falch-Larsen <even.falch-larsen@nordicsemi.no>
Co-authored-by: Uma Praseeda <uma.praseeda@nordicsemi.no>