lib: lte_lc: Initialize semaphore in init function
When CONFIG_LTE_AUTO_INIT_AND_CONNECT is disabled and modem is switched
to online mode using AT commands the link semaphore is used before it's
initialized and crashes the application. This is not the usual use case,
but we ran into this with a shell application where user can easily mix
LTE LC and AT command usage and the fix is simple.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>