ci: allow publishing to multiple servers
Update publishing instructions to file to use "publish2"; this command
gets an extra argument providing a location name to use for publishing.

The location must match the names used in the server side script, which
currently are "dev" for the development server, and "main" for the
production server. PRs are left published to the development server, and
switch "master" and other branches (currently disabled in the workflow)
to push to the production server.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>