doc: move versions.txt generation to own file
This fixes an issue after running `ninja clean-nrf` followed by `ninja
nrf`. Since `cmake` is not invoked, the global `configure_file` was not
being run and would not regenerate the pre-requisite versions.txt file.

The version generation was moved into another "cmake" file and the
versions.txt is now an output of  add_custom_command.
versions.txt will be re-generated whenever it is missing or one of its
input is modified.

The nrf-content depends on versions.txt ensuring that it will always be
up-to date before nrf-content target is executed.

JIRA: NCSDK-7514

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>
Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>