doc: move versions.txt generation to own file
This fixes an issue after running `ninja clean-nrf` followed by `ninja
nrf`. Since `cmake` is not invoked, the global `configure_file` was not
being run and would not regenerate the pre-requisite versions.txt file.

The version generation was moved into another "cmake" file and imported
as a dependency of the nrf-content target, forcing it to always run on
each build.

JIRA: NCSDK-7514

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>