spm: support for nRF5340 peripherals in non-secure applications
The current SPM only supports peripherals that are present on the
nRF9160. While there are some overlap, many of the nRF5340 peripherals
are not supported. These are added to spm.c and Kconfig with this PR.

In addition DTS-generated macros are now used for peripherals that
has this.

Tested by building hello_world for nrf5340dk_nrf5340_cpuappns and
checking that the peripherals are actually found by SPM.

Signed-off-by: Carl Richard Steen Fosse <carl.fosse@nordicsemi.no>