Bluetooth: Mesh: Light Ctrl srv: Perform regulator step in interrupt
Moves the light control server's regulator timeout to a k_timer, instead
of a k_delayed_work. This ensures that the regulator runs at a regular
interval without being encumbered by slow threads.

The callback to the lightness server is deferred to a k_work item, and
may still be subject to excessive delays in other threads.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>