net: lib: nrf_cloud: Let control and data messages share ID counter
The MQTT specification states that there can be only one unacknowledged
message with a given message ID in a connection. In the nRF Cloud
library, the data and control messages have had separate message ID
counters, which can lead to the same message ID being used at the same
time, resulting in the last arriving of them to be dropped by the
broker. This patch fixes this issue by introducing a shared counter.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>