net: lib: azure_fota: Add job ID handling to avoid repeated update
Add job handling of job ID to avoid repeatedly downloading the same
firmware image.

In certain situations, the device can get into a loop
where it concinues to download the same firmware image over and
over again without being able to apply it, spending large amounts of
power and data traffic.
This patch prevents that from happening by reporting the job ID that
was last processed and comparing it to the ID of any incoming firmware
update.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>