samples: connectedhomeip: include CHIP as an NCS module
So far, CHIP has been included in an NCS-based project as
an out-of-tree Zephyr module to assure that it is possible
to switch between NCS-supplied and standalone CHIP version.

Now that https://github.com/nrfconnect/sdk-zephyr/pull/436
has been merged, it is possible to override a module which
is already on the list. Consequently, there is no harm in
including the NCS-supplied CHIP automatically since
applications are able to use an upstream CHIP revision
if needed.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>