applications: asset_tracker_v2: Send config if received config is empty
Add an event that notifies the data module if the received desired
configuration is empty. The data module acts on this event by sending
the currently applied device configuration. This is either the device's
default configuration or the lastly applied config retrieved from flash.

The device should report its configuration after boot and the
asset_tracker_v2 has a test specifically for this.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>