applications: asset_tracker_v2: subscribe to rejected get requests
This is a follow up to #3881.

Subscribe to the $aws/things/thingName/shadow/get/reject topic
so that the application can handle the case when the shadow does not
exist, which happens for new devices which have never connected
before and use AWS IoT just-in-time-provisioning to set up the AWS IoT
Thing after the first request.

The application tries to get the shadow, which is rejected, and this
will trigger the reporting of the cfg property with the default
values.

Signed-off-by: Markus Tacker <Markus.Tacker@NordicSemi.no>
Co-authored-by: Simen S. Røstad <simen.rostad@nordicsemi.no>