drivers: serial: lpuart: Fix handling of corner cases
It may happen that receiver was reset and transmitter is finishing
previous transfer or is in reset state and request pin is not
driven by anyone.

Lpuart was not handling correctly the case when after initialization,
but before enabling RX interrupt, a packet was received. In that
case, after enabling RX interrupt fifo was read but freed buffer
was not feed to the lpuart. Fixed that case by adding RX buffer
feeding in RX interrupt enabling function.

In case transmitter was in reset state, RX pin floated leading to
unexpected gpio interrupts. Updated driver to support handling
those cases.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>