net: lib: aws_fota: Fix issues with reciving multiple notify-next
When entering PSM the network will cache TCP packets. This lead to
multiple notify-next messages being sent when the device was in PSM as
the MQTT ack was never recived so the server would do retrasmission.

To combat this we check what state we are in when we receive a
notify-next topic so that we won't accept it and break the download.

Addition to this we add a timeout on accepted so that things running in
the same context won't be blocked forever.

Ref: NCSDK-8251

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>