net: lib: aws_iot: Add own event for FOTA errors.
Add AWS_IOT_EVT_FOTA_ERROR event that is used to propagate FOTA-related
errors to the application. This is to distinguish between
AWS_IOT irrecoverable errors and FOTA errors. If FOTA fails it is not
necessarily a reason for a reboot, and it should be up to the
application if a reboot should occur.

Add the corresponding evt CLOUD_EVT_FOTA_ERROR for the generic
cloud API.

Update samples that uses AWS IoT and cloud API to use the new events.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>