net: lib: azure_fota: Fix issues with reporting of 'error' status
This fixes an issue where a reported error for a FOTA job is
overwritten after reboot.
In addition, a check of internal state is added to avoid wrongly
reporting 'error' in case a download is in progress when a 'reported'
object is received with FOTA status set to 'downloading'.

Fixes NCSDK-8318

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>