tests: fprotect: Add tests for read/write protection
Enable HUK loading and test that it is unreadable.
Also, add a fatal error handler so the test passes after each fatal
error.
Add a dummy_huk.hex that must be part of the hex file. It only contains
the magic number, no actual key (the 0xFFs become the key).

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>