net: lib: aws_fota: Clear MQTT buffer when message is dropped
When a device receives multiple notify-next messages, only
the first should be acted upon. The subsequent notify-next messages
are not processed by the library, but the payload must still be
read out so that the MQTT library can correctly receive the
next incoming messages. The message should still be marked as
an AWS FOTA message to avoid it being forwarded to the application.
This is done by returning 0 from the AWS FOTA event handler.

This patch fixes an issue where duplicate  job messages were not
marked as AWS FOTA messages and makes sure that payload is read out
for these messages.

Fixes NCSDK-8251

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>