applications: asset_tracker_v2: Purge message queues if enqueue fails.
Purge message queue if k_msgq_put() returns an error.
This is to ensure that the calling module can continue
to process events.

This fixes the issue where a module with a full message queue
would never handle the incoming shutdown request event from the utility
module after an error.

Closes CIA-240

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>