doc: known issues: fix font size
We are using definition lists to display each known issue.
The RST processing tries to be smart by picking a font size
depending on how much content each entry in the definition
list has. However, we want the same font size for every
definition list on this page - therefore, adapting the script
to insert class="simple" for all definition lists (not only
the ones that the processing picks).

Signed-off-by: Ruth Fuchss <ruth.fuchss@nordicsemi.no>