applications: asset_tracker_v2: Fix data types passed between modules.
When passing data between modules that sample data we should copy the
content on the event instead of passing pointers to the data.
By using pointers we don't know if the memory pointed to is valid at the
point of use (or being altered by another context).

This change imposes an increase in the overall stack size needed by the
modem data specific ringbuffers. Because of this we decrease the number
of ringbuffer entries and increase stack size of modules including modem
specific data.

This patch also increases the type size of cellID. This was previous
converted from a double (from the modem info module) to an
unsigned 16 bit. Unsigned 16 does not have a sufficient maximum
representative range to store the value for the cell ID.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>