samples: nrf9160: mqtt_simple: Randomize client ID by default
Currently, the Client ID the MQTT client connected with was statically
defined with the Kconfig. Because of this, every user of the sample
which did not customize the Client ID was connecting with the same,
default value. This resulted in some unexpected behaviour, when the MQTT
server disconnected the existing connection when a new client with the
same Client ID connected to the server.

Fix this, by using a unique Client ID in the sample. For nRF91 based
devices, the Client ID is generated using its IMEI number, other devices
will use the board name and a 10-digit random number. The option to set
the Client ID with Kconfig is still available - if set, it will override
the default value.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>