pcd: don't set secure attribute on PCD RAM region
There is no reason to set the secure attribute on the RAM
region used for IPC in the PCD module. This because the
region is write-protected, and there is no sensitive
information placed in the region.

Not setting the secure attribute allows the user of the
system to configure the network core as non-secure and
still leverage the PCD functionality to have upgrade
capabilities of the network core.

Ref: NCSIDB-409

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>