lib: lte_lc: Use fallback when only one LTE mode is enabled
If both LTE-M and NB-IoT are enabled at the same time, the "manual"
fallback mechanism should not be used. Instead, the LTE preference
configuration should be used to control which technology is used to
connect to the network. Using the modem's logic to control the
switching of mode is likely to be more efficient both in terms
og time to connection establishment and energy consumption.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>