applications: serial_lte_modem Protect UART RX in data mode
It seems that uart_rx_disable() doesn't have immediately effect
that if there is more date to reveive, the driver is disabled
after all data is received.

This create a problem for data mode that <size_limit> triggered
transmit could or could not include new data from Serial driver
callback (higher priority), while the original design is should
not include.

Add a new rule, if a transmit is scheduled (triggered by any of
<size_limit>, <time_limit> or RX_RDY), SLM drops data received
after this point, until transmit is done and UART RX re-enabled.
Also explicitly leave a line in the log of data dropping.

Typical case, <size_limit> set as 1024 and MCU send 1025 byte.

Signed-off-by: Jun Qing Zou <jun.qing.zou@nordicsemi.no>