applications: asset_tracker_v2: Disable/enable ADXL362 trigger on events
This patch adds events to the application module that lets it decide
when it wants movement to trigger the next sample request.

Prior to this patch the accelerometer ringbuffer in the data module
would get saturated with accelerometer data from the sensor module
after the initial movement trigger. If the device was under constant
movement over the set threshold each cloud update would include
multiple instances of accelerometer data in the batch data update.

Changes included in this patch:
- Add function to external sensors API to enable/disable callbacks
  from ADXL362 driver.
- Add configurable option to decide if callbacks from the ADXL362 driver
  should be enabled during initialization of external sensor API.
- Add events to application module that signals the sensor module when
  it wants movement to trigger the next sample request.
- Set timers and send the new sensor events with private APIs in the
  application module.

Closes CIA-231 and CIA-278

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>