applications: asset_tracker_v2: Add Kconfigs to optionally store data.
Add options that decide if data of the respective cloud data type
received from other modules should be stored to its respective
ringbuffer.

Disable storage of accelerometer data by default. The data that we
get from the accelerometer does not currently provide any value to the
end-user and should be removed for the time being. This is done for
power and data saving purposes.

There are plans to improve the way data from the accelerometer is
handled in the firmware and at that point storage of accelerometer data
will be reinstated in a different form than just single sample XYZ
coordinates.

Closes CIA-281

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>