application: asset_tracker_v2: Add function to filter duplicate messages
Add function that keeps track of the version number of incoming
message from the AWS IoT broker. If a message has already been handled
by the application, it is ignored.

Messages from the AWS IoT shadow contain a version number that
increments for each new message published to the device.

Closes CIA-288

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>