pcd: Split PCD config into PCD_APP and PCD_NET
The PCD Kconfig option includes code for both the application processor
and network processor.  This brings in code that does not make sense to
run on both processors as well as dependencies that are not needed.
This change creates two separate Kconfig options: PCD_APP includes the
code for the application processor only and PCD_NET includes the code
for the network processor only.  The PCD option is retained for
backward compatability.  The Kconfig option PCD_USE_CONSTANTS was added
to allow configuration of the constants used by the pcd code via
Kconfig.  These changes allow PCD code to be used by more custom boot
loaders.

Signed-off-by: Denis Bohm <denis@fireflydesign.com>
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>