applications: asset_tracker_v2: Improve module shutdown handling
This patch adds functionality to the application that enables it to
keep track of what modules that supports shutdown.

This is done by maintaining a list in the modules common library of
modules that supports this feature. This is done by passing in a flag
when starting the module. Modules will acknowledge that they have shut
down by sending an event to the utility module including
their module ID. The utility module will then register that the
respective module has shut down and when all modules have acknowledged
the shutdown, the application reboots.

This patch includes:

- Add new macro to lib modules common to send shutdown ACK events.
- Introduce module ID that is used to identify modules.
- Keep track of an internal linked list of modules that support
  shutdown.
- Add new API to lib modules common to register modules that have
  performed a shutdown.

Closes CIA-244

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>