dfu: Correct calculation of last erased page offset
When loading stream write progress the offset of the last erased page
can be set incorrectly one page ahead whenever the write ended just
after a page boundary. This commit fixes that by subtracting one from
the calculation. It also checks to see if progress was stored with
zero bytes written, in which case it sets the last erased page start
offset to its default value.

Ref: NCSDK-8901

Signed-off-by: Jonathan Nilsen <Jonathan.Nilsen@nordicsemi.no>