.github: add workflows/oss-history.yml
Currently this just checks the zephyr history for 'rebasability'.
This is useful for making sure our downstream history is OK and can
be rewritten for release time.

We will add support for other repositories as we go. This will make it
easier to cut releases since we know our trees are clean. Eventually
we should also save these rewritten histories to be able to switch to
them any time.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>