partition_manager: don't append pm report target to rom_report target
The two reports (rom_report and partition_manager_report) does
not have anything to do with each other. Hence it makes little
sense that we append pm report to the rom_report target.

Move the declaration of 'partition_manager_report' to the
partition_manager.cmake file to avoid having to perform
genexp operations to function before the 'partition_manager'
target has been created.

Also remove support for the 'quiet' flag in the report script, and
update the 'input' argument in that script to require arguments.

Ref: NCSIDB-278

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>