net: lib: nrf_cloud: add config options for client id
Add config options for setting the MQTT client ID.
For mfw 1.3, the preferred ID is the modem internal UUID.
This can only be obtained by parsing it out of an attestation token.
Also include an option to set the client ID at runtime.
This is useful if the ID is stored somewhere else on the device
and isn't known to the nrf_cloud lib.
Moved the compile time option to kconfig.
IMEI-prefixed option remains and is the default for now.

Signed-off-by: Justin Morton <justin.morton@nordicsemi.no>