lib: lwm2m_carrier: Depend on FP_HARDABI in kconfig
Adds dependency on FP_HARDABI for the LwM2M carrier lib, which is only
compiled for hard-float. Adds CONFIG_FPU=y and CONFIG_FP_HARDFLOAT=y in
all samples and overlays using this library to ensure that it gets
included.

This is a followup to nrfconnect/sdk-nrfxlib#460, which removes
imply FPU for the NRF_MODEM_LIB, causing build failures in samples
relying on this imply to enable the FPU for the LwM2M carrier.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>