applications: asset_tracker: Explicitly enable CONFIG_FPU
This is a followup to nrfconnect/sdk-nrfxlib#460, which removes
imply FPU for the NRF_MODEM_LIB, causing build failures in samples
relying on this imply to enable the FPU.

Without CONFIG_FPU, the BME680_BSEC library linking fails, as it's only
compiled for hard-float. Explicitly setting this in prj.conf for the
asset_tracker preserves the behavior of the old imply made by nRF Modem
lib.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>