lib: nrf_modem: Fix duplex operation of the offloaded sockets
Due to the recent changes in the socket layer in upstream Zephyr,
regarding an extra mutex protection, we're not longer able to
transmit and receive in parallel from separate threads.

The upstream socket implementation handled this problem with `condvar`
kernel object, but the implmenetation is specific to the native socket
implementation and is not appicable for the offloaded sockets. Therefore
handle the issue in another way, by releasing the lock for the blocking
nrf_recvfrom calls, which prevented the other threads from sending.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>