applications: asset_tracker: Add Predicted GPS Support
When initializing PGPS, give it the address of the
MCUboot secondary scratchpad area to store predictions in.

Pass AGPS request to PGPS module as well if PGPS is enabled.
Add pgps callback for significant PGPS events.
Add PGPS work function for injecting prediction to GPS unit.
Enable PGPS in thingy91 and 9160dk builds (for now).
Enable using AGPS + PGPS as well as each separately.

Add overlay-pgps.conf for building with P-GPS and not A-GPS,
and overlay-agps-pgps.conf for building with both.

Add reference to overlay files in documentation.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>