samples: nrf9160: agps: Add PGPS support
Using the overlay-pgps.conf or overlay-agps-pgps.conf
Kconfig overlays, enable PGPS features. They
include settings for MCUBOOT, which is only needed
to define the symbol and address of
PM_MCUBOOT_SECONDARY_ADDRESS. Pass that address
to the PGPS init function as the place to store
predictions.

Add conditionally-compiled hooks to main.c
so that PGPS is initialized, data received from
cloud is processed, a relevant prediction can be
selected, and that prediction injected to the
modem instead of getting it on demand from the
cloud via AGPS.

Make this work with AGPS alone, PGPS alone,
or both together.

Build with west this way:
west build -p auto -b nrf9160dk_nrf9160ns \
-d build_9160dk_agps samples/nrf9160/agps \
-- -DOVERLAY_CONFIG=overlay-pgps.conf

Add reference to overlay files to documentation.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>