lib: lte_lc: Always parse current cell when available
Current cell information was only parsed if there was also neighbor
cell information present in the %NCELLMEAS response.
This change parses current cell information when it's available and
populates the event struct accordingly. The current cell ID can
be used to check if current cell information is valid in an event.

Test case is updated to verify the changed behavior.

Closes CIA-309

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>