mpsl: Make signaling semaphore binary
Semaphore's count value can increase in an uncontrolled manner during
certain multiprotocol scenarios. As an example take Bluetooth scanning
and IEEE 802.15.4 RX. The IEEE 802.15.4 driver tries to obtain a
timeslot. The request is denied due to scanning activity. In the signal
handler the driver requests the timeslot again. This chain of events
creates a busy loop during which the best case scenario is one semaphore
taken and one semaphore given. Any other signal that shall be handled in
the low priority handler bumps the counter. The counting semaphore must
be then unwound to return control to other threads. This can introduce
performance issues and even block other threads entirely.
Changing the counting semaphore to binary resolves the issue as there is
no unwinding required.

Signed-off-by: Pawel Kwiek <pawel.kwiek@nordicsemi.no>