caf: ble_state: Fix build issues for nRF5340
Building for nRF5340 is split into 2 images, App and Net mcu.
BT_CTLR is a part of Network mcu. If Softdevice is chosen as LL,
TX_POWER must be set manually by App mcu. Because
CAF_BLE_STATE_TX_PWR depends on CTLR it was impossible to determine
proper value.

Workaround that by setting TX_PWR to 0 if CTLR is not defined.

Jira: NCSDK-9476

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>