Bluetooth: Mesh: Permit transition time = NULL
Pull request #3827 changed the Generic OnOff's set message transition
time parameter to NULL for certain scenarios. This change is propagated
to higher level models, which forward the OnOff set message in some way.

To make this behavior consistent, this change makes the transition
pointer NULLable in all APIs, and makes all server models follow the
same pattern as the Generic OnOff model, so that when no transition time
parameters are provided (either in the message, or through DTT), the
transition pointer is NULL.

Updates documentation and samples, and replaces some code that jumps
through hoops to ensure that transition isn't NULL. See handling of
Generic Level's Move set in higher level models, in particular.

Replaces model_transition_is_active with a public function that returns
the total transition time.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>