samples: profiler: Add qemu_cortex_m3 to excluded platforms
This is a follow-up to commit 75f093bce97f0e1d7f658934a67f0ceefdfc5625.

This sample uses the Nordic profiler subsystem which bases on RTT,
so it cannot be built on any platform not selecting HAS_SEGGER_RTT
(unfortunately, unsupported platforms cannot be filtered out by
using the filter property, because the Kconfig dependency error occurs
when the sample is configured in cmake).

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>