applications: asset_tracker_v2: Increase default PSM timers.
Increase PSM TAU timer to 320 hours and PSM Active time to 6 seconds.
320 hours is just an arbitrary high number that will most likely be
higher than the publication interval for most applications. 6 seconds
active time is to give some slack to prevent the device from going into
PSM too quickly. A common occurrance when a too small active time is
given is that the MQTT broker may not be fast enough to respond to QoS 1
messages.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>