samples: connectedhomeip: fix hangs on sys_rand32_get()
There is a known issue in the cryptocell library causing
that an application may hang on a call to sys_rand32_get()
or underlying functions. The issue is supposed to be fixed
in the next cryptocell library version, but for now provide
a workaround using the non-cryptographically secure
Xoroshiro algorithm for random number generation.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>