Bluetooth: Mesh: Use instant transition time for the Light LC regulator
When running the regulator, the light control server would attempt to
guess the interval until the next regulator step would be able to
execute, and use this as the transition time for the lightness
adjustments. This is not accurate, and generally introduces a lot of
overhead for applications that run their transitions using CPU timers.

Replace this behavior with instant transition time to relieve the
application of this overhead, as it's not accurate anyway.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>