net: lib: nrf_cloud: preemptive update recovery
Fix bugs in handling preemptive updates.

If the first update fails, we did not increase the number to replace
expired updates as needed.  Further, we would not retry the request
until the threshold has been expired again.

This fix retries the request on each call, and grows the request size
as needed. It also turns off PGPS_MARGIN_SEC -- nonzero values allows
expired ephemerides to be selected, which then defers updating to new
ones.

Jira: NCSDK-9960

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>