net: fota_download: Add state variable to check if it is downloading
Due to the callback nature of download client and the delayed scheduling
of a restart from offset it is possible to call `fota_download_start`
while something else is downloading which causes an error.

This commit adds a state variable which we can check to make sure that
we don't start a new download while another download is still in
progress.

Ref. NCSDK-8541

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>