lib: multicell_location: Add options for socket timeouts
Add Kconfig options to set timeouts for sending and receiving.
This addresses isses that have been observed where the connection
is not closed and the library never starts to parse the received
data.
Now the library will attempt to parse the data in the case of
timeout, as it may have received all the required payload even
though a timout occurs.

CIA-333

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>