lib: sms: Acknowledge unsupported/invalid SMS.
When received SMS is malformed or using features that we don't support,
it was not acknowledged with AT+CNMA=1. This caused SMS service
between the device and LTE network to get stuck, i.e.,
no further messages could be sent or received.
An example is a message with UCS2 data encoding with an emoji or
Asian characters. This was not supported and just ignored without
acknowledging it. We'll continue to ignore the message content but
will acknowledge it to keep SMS service working.
Improved error logging related to UCS2 encoded message.
Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>