fprotect: Fix bug when multiple boot stages are present (ACL only)
If a bootloader uses fprotect, the spent ACL region weren't accounted
for if e.g. the app tried to protect something.

Also add regression tests for this bug in the negative fprotect tests.
Refactor the test to test both the areas the immutable bootloader should
protect, and areas protected by the test itself.

Ref: NCSDK-10164

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>