fprotect: Add fprotect_is_protected API (NRF_ACL only)
To check whether an area has been protected with fprotect.
This is useful because the device will through an exception if a
protected area is read or written.

Also, remove #if defined(CONFIG_FPROTECT_ENABLE_NO_ACCESS) from the
header file since the coding standard says that functions should not
be ifdefed out in header files, only in source files.

Ref: NCSDK-10192

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>