hw_unique_key: Improve HUK loading in the immutable bootloader
Add a separate config CONFIG_HW_UNIQUE_KEY_LOAD which doesn't enable
as many other configs, to give access only to the KDR loading API.

Modify the immutable bootloader to allow a single boot without
a HUK present. The bootloader writes the last word of the HUK flash
page to 0, to signify that it should fail if a HUK is not present.

This allows the application to write a random HUK. Random number
generation requires quite a bit of code, making it infeasible to add
to the bootloader. Therefore, this change removes the requirement to
either
 - Flash random_hw_unique_key without the bootloader, boot it, and
   then flash the bootloader and app, or
 - Generate a key off chip and programming it together with the
   bootloader and app.

Ref: NCSDK-9034

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>