doc: known_issues: Add section about FMFU and newest J-Link
There is an issue with the newer J-Link versions which causes the UART
communication to break when sending the modem firmware data using the
SMP protocol.

This affects every version of the FMFU SMP Server sample. Using J-Link
firmware on the debugger chip from 6.88a fixes this issue.

Ref. NCSDK-9441

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>