zigbee: Improve ZBOSS timer accuracy
Since the workaround for ZOI-686 has been accidentally removed from the
ZBOSS, there is a higher demand on the ZBOSS timer accuracy.
This PR reduces the period of the Zephyr counter, used in OSIF to reduce
an accumulated error, caused by mismatch between expected and measured
periods (15.34 ms vs 15.36 ms).

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>