net: lib: azure_iot_hub: Fix too small property bag buffer
When DPS is enabled, the device ID is used as the DPS registration
ID. The registration ID is passed back to the device as a property
bag, which means that a property bag value can be as long as the
device ID in the worst case.

This patch asserts that the property bag buffer is at least as
large as the device ID buffer.

In addition, the default value of the property bag element size
is increased.

NCSDK-10175

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>