samples: spm: Disable UART rx pin to get low power
When rx-pin property is not set for given uart instance,
receiver is not started and uart is disabled after each
TX transfer. This results in UARTE not drawing any power
when not in use for console/logging.

Delete rx-pin property for uart0 and uart1 in spm. After
that change spm will print initial status but will jump
to the NS application with uart is the lowest power state
(disabled).

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>