bootloader: Change slot alignment to avoid problems in some cases
When the mcuboot pad was not needed, the start of S1 was not properly
aligned, which could lead to it starting in the middle of a block.

We cannot align on both end and start, so to ensure that the start
of S0, start of S1 and end of S1 are all aligned, we now use the
new 'align_next' partition manager directive.

Ref: NCSDK-10171, NCSDK-10209

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>