spm: Fix size calculation of NSC region
The existing calculation is only correct for 32 and 64 bytes.
The calculation was doing division instead of log2.
Our secure services are below 32 bytes, so this hasn't affected us.
From 64 bytes to 256 bytes, the size configured in HW will be larger
than needed, and above 256 bytes, the HW configuration will be invalid.

Ref: NCSDK-5156

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>