manifest: get openthread from zephyrproject
We've been carrying a separate openthread fork in the 'nrfconnect'
GitHub organization during the NCS v1.6.0 development cycle. However,
there's no need to be doing that anymore.

After looking into it, I noticed we are using a version whose code is
already available in the 'zephyrproject-rtos' GitHub fork.

It's not what zephyr/west.yml is currently pointing to, so we can't
just pull openthread from the allowlist of projects we get from that
file. But we can just point at the zephyrproject fork instead of our
own.

Note that the 'thread-reference-20200818-ncs1-rc3' tag we have right
now is not the exact same commit as
f460532d4afa5d49feba241e5dc31c56123d31a8, but there is an empty diff
between them, so they are equivalent repositories.

Using this SHA from zephyrproject avoids the need to release NCS
v1.6.0 with an additional layer of openthread forking, which is
something we've never done before and is therefore worth avoiding.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>