samples: nrf9160: gps: support circuitdojo_feather_nrf9160
This platform has only external antenna, so narrow down possible Kconfig
options to reduce user confusion.

Set COEX0 pin to 1 when used with GPS frequencies, as the external
antenna signal goes through LNA, which has to be enabled.

MAPGPIOs are not used on this platform, so do not configure them.

Signed-off-by: Marcin Niestroj <m.niestroj@emb.dev>