partition_manager: add offset from region start
A region has a base address and currently partition manager
only outputs the absolute address to the partitions, not
any offsets from the base of the region it is located in.

This is problematic for several APIs that operate on offsets.
Working with offsets also mimics the devicetree model
more closely.

To work better with these APIs, and to be more aligned with
devicetree we add the _OFFSET define in the output of
partition manager.

In addition we change flash_map_pm.h to provide this value
to the 'FLASH_AREA_OFFSET' macro, which is what users
would expect in the first place.

Ref: NCSDK-9797
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>