connectivity_bridge: fix malfunction by replacing the UART driver
An improved UART driver for nRF52840 boards is provided which
offers very low latency, simple to use API and is more reliable
than the original driver. The new driver is integrated in the
connectivity_bridge application, to address reliability issues
with the original driver.

https://devzone.nordicsemi.com/f/nordic-q-a/67372/connectivity_bridge-malfunctions-when-ble-client-disconnects-and-reconnects

Fix malfunction in connectivity_bridge caused by the UART driver:
* connectivity_bridge randomly stops forwarding data from UART
to BLE and USB clients. The issue resolves on its own within 5
minutes, or when the board is restarted. The issue usually
appears a few minutes after boot. In one test scenario, we had
a very active bi-directional communication between nRF9160 and
a BLE device. In another test we were sending AT commands
automatically from the BLE client, to monitor the modem status.
In a third test, we were sending the AT commands manually over
the USB terminal. During most of our tests, a USB terminal was
also attached to monitor the communication. When the issue
appears, both USB terminal and BLE client are affected and
stop receiving data for about 5 minutes.

This patch resolves the aforementioned issues, by offering an
alternative UART miniport driver for nRF52840. The driver is
well tested and reliable. EasyDMA is enabled by default.
Optionally the driver can be switched to interrupt or polling.
All connectivity bridge features and power saving are preserved.
Board configuration is used e.g. enabled, pins, baudrate, and
hardware flow control. The driver API is very simple to use.
It is designed to provide very low latency and efficient
CPU usage.

When CONFIG_NRFX_UARTE is enabled, connectivity_bridge uses the
original UART driver. When disabled, the UART miniport is
enabled as a fall-back. It is recommended to disable all
CONFIG_NRFX_UARTE* and CONFIG_UART* features, see prj.conf.
CONFIG_DYNAMIC_INTERRUPTS, and
CONFIG_DYNAMIC_DIRECT_INTERRUPTS must be enabled.

Added features for nRF9160dk: Thingy:91 like behaviour:
Data received over UART0 and UART1 is sent over TX on the same
UART: data received from nRF9160 is forwarded to VCOM1, as well
as the BLE client. Data received from VCOM1 or BLE is forwarded
to nRF9160. Pin configuration should be configured accordingly
in the board files. UART power saving is disabled for this board,
since nRF5240 is not aware when a USB terminal is attached to
the board controller. UARTs are cross-wired: nRF9160 connects to
UART0.RX and UART1.TX, while VCOM1 connects to
UART0.TX and UART1.RX.

TODO:
While the suggested changes to connectivity_bridge are well tested
and reliable, it would be best to integrate the UART miniport
driver into Zephyr or make it an optional NRF component, which can
be selected instead of the original driver. It is suggested to
create an option CONFIG_UART_MINIPORT_ENABLE, that automatically
enables and makes the driver available for use in any application.

Signed-off-by: Georgi Valkov <gvalkov@abv.bg>