samples: nrf9160: gps: fix IP address and socket connect handling with SUPL
The used SUPL server IP address was printed incorrectly before trying to
connect to the server.

The implementation tried to use an IPv6 address, although only IPv4 is
currently supported.

If connecting to the server failed, the function still returned 0 and
caused the SUPL library to try to send data using the socket.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>