partition manager: configure MBR partition for nRF5280 dongle
nRF52840 dongle reserves the first 4KiB of flash for the
MBR. Partition manager does not reflect this fact and
MCUBoot is built with a wrong base address - 0. As a result,
any sample built with CONFIG_BOOTLOADER_MCUBOOT=y does not
run correctly on the dongle.

Add a new "mbr" partition when an application is built for
the nrf52840dnogle_nrf52840 target.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>