samples: use kconfig to include child images on other core on nrf53
Leverage kconfig to decide what child image should be included
on the other core.

Prior to this many kconfig option checks has to be performed
in cmake in order to decide what child images to include.

Also, it is not possible for out of tree users to include
arbitrary samples as child images for the different cores.

This solution fixes both of these issues by leveraging kconfig
to decide what sample to include as a child image at the other
core.

The design goal of this solution is as follows:
- Any child image which was automatically included in the old solution
  should be automatically included in this solution.
- It should be possible to opt-out of automatic inclusion.
- It should be possible to opt-in for automatic inclusion of a sample
  in configurations other than those that will automatically include
  them.
- If a sample is automatically included the options for child image
  names and paths should be removed to avoid confusion.
- A user should be able to pass an arbitrary child image name and path
  if no child image is automatically included.

Ref: NCSDK-6992

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>