samples: use kconfig to include child images on other core on nrf53
Leverage kconfig to decide what child image should be included
on the network core.

Prior to this many kconfig option checks had to be performed
in cmake in order to decide what child images to include.

Also, it is not possible for out of tree users to include
arbitrary samples as child images for the different cores.

This solution fixes both of these issues by leveraging kconfig
to decide what sample to include as a child image the network
core.

The design goal of this solution is as follows:
- Any child image which was automatically included in the old solution
  should be automatically included in this solution.
- It should be possible to opt-out of automatic inclusion.
- A user opting-out of the automatic inclusion should then use
  add_child_image() to add his custom image.

Ref: NCSDK-6992

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>