subsystem: adding event manager storage layer
Depending on the application, it can be useful to store transient
event into a specified storage technology to be able to replay
events that were not consumed by the application.

This may happen if the board ran out of battery or if the application
crashed during the processing of the event (or any other reason).

This commit adds a new storage layer to the event manager, similar
to the storage layer of the Zephyr settings API.
For now, only NVS is supported, but littlefs is planned. The
configuration file + a basic source file are already added to
that commit.

Each time an event is submitted, it is first sent to the storage
layer. This layer will give a unique id to the event and will
store the event. Once the storage of the event is done (and
regardless of the result of that operation), the event is truly
sent to the manager which will then call all the subscribers.

Each time an event is consumed, the storage layer removes it from
the storage.

When the board boots (or when event_manager_init() is called to be
more precise), the event storage layer is initialized.
Once that initialization is done, it will load each non-consumed event
present in the storage.
Each of these events will then be automatically re-submitted to the
application.
If the event is properly consumed, it will be removed from the storage.

Know issues:
-----------------------------
- Dynamic sized event are not supported yet

Roadmap:
-----------------------------
- Adding more storage option
- Supporting dynamically sized events
- To prevent reading/writing too frequently we may add a flag in the
event header that will tell the storage layer to ignore an event.
For example a "button_pressed_event" can be forgotten without much
risk. However, "machine_about_to_blow_event" must be stored and
replayed if need be.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>