applications: asset_tracker_v2: Add new "nod" run-time configuration
Add support for a new run-time configuration that enables
sample requests for GNSS and neighbor cell data to be enabled/disabled
at run-time.

At JSON protocol level, the configuration is called "nod" which is short
for No Data. This is a list that includes data types that are not
going to be included by the application module when sample requests are
sent out to other modules. This enables the user to configure via the
web interface the source of location-related data that is sampled on
the device. If the list is empty both GNSS and neighbor cell data
is sampled.

In the application, this list translates to boolean variables that are
used to toggle the inclusion of the corresponding data type in the
`APP_EVT_DATA_GET` event.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>