subsystem: Adding the concept of important event.
To prevent saving all the events to the non-volatile storage,
the concept of important event is introduced.
Only important event will be stored to the non-volatile storage
and replayed upon reboot if not properly consummed.

This has 2 main pros:
- It will protect the flash lifetime by greatly reducing the
number of read/write operations on it

- It will only replay events that are really needed, and will
prevent replaying an event in the middle of an event chain.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>