net: lib: download_client: fix expected HTTP status
When not using TLS, the library made an HTTP range request
when requesting a resource from the server (requesting all the bytes).
This led the server to reply 206 but send the whole resource
in one HTTP message. This used to work, but a recently introduced
change expected the HTTP status in the response to be 200
when not using TLS, thus it started failing with EBADMSG.

This patch lets the library omit the Range header in the request,
(and expect 200) when downloading without TLS,
unless downloading from an offset.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>