cmake: add cache entry for child image config files
When using child image configuration files, then those were not added to
the CMake cache.

This made the config file knowledge unavailable to other parts of the
system, such as mcuboot.

In a multi image scenario, mcuboot will use the path information from
mcuboot_CONF_FILE to lookup the private key for signing.

Adding the <aci_name>_CONF_FILE to the CMake cache when present ensures
that other parts of system will be able to lookup the need information.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>