manifest: MPSL with public Coexistence API
Update MPSL version in nrfxlib to expose Generic Coexistence API
intended to be used by protocol stacks depending on MPSL: nRF 802.15.4
Radio Driver, SoftDevice Controller.
As a side effect, SoftDevice Controller revision is updated as well.

This is initial step towards Generic Coexistence API usage. After this
PR is merged, it will be followed with example implementation of new API
and changes in 802.15.4 stack to use this example implementation as PoC.

Signed-off-by: Hubert Miś <hubert.mis@nordicsemi.no>