Bluetooth: Mesh: Lightness: Add bt_mesh_ prefix to internal API
Prefixes lightness internal functions with global linkage with bt_mesh_
to avoid breaking out of the mesh stack namespace.

Even though this header isn't exposed to the application, the linker
will still treat the non-static functions as global symbols, and they
may conflict with application functions of the same name.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>