net: lib: nrf_cloud: Add option to disable auto-request of P-GPS data
Add option that enables the user to disable auto requesting of P-GPS
data after initialization of the nRF Cloud PGPS library.

This option is now disabled in the Asset Tracker v2, reason:

Prior to this, requesting a full set of P-GPS data would be attempted
twice if `nrf_cloud_pgps_init` and `nrf_cloud_pgps_notify_prediction`
were called in close succession. In the application we call
`nrf_cloud_pgps_notify_prediction` upon every request from the modem
that it needs assistance data. This request often occurs at the same
time as the intialization of the aforementioned library.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>