lib: pdn: update KConfig dependency and default for SYS_INIT
Add a depends on NRF_MODEM_LIB_SYS_INIT to PDN_SYS_INIT.
The dependency is necessary to be able to send the
AT commands that override the default PDP configuration,
and control the PCO/ePCO setting used by PDN connections.

In addition, we default this option to `y` when the
NRF_MODEM_LIB_SYS_INIT is set, for convenience.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>