net: lib: nrf_cloud: add GNSS API support for nRFCloud A-GPS and P-GPS
Added support for writing assistance data to GNSS using the new GNSS
API. If GPS driver is not enabled and no GNSS socket descriptor has
been given, it's assumed that the GNSS API is used.

Removed Kconfig dependency to the GPS driver from both A-GPS and P-GPS.
It's should not be mandatory for the application to enable the GPS
driver. The library does use gps.h, but that's always present
regardless of the NRF9160_GPS configuration option.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>