lib: lte_link_control: multiple registered LTE event handlers
There is a need to register for LTE events in multiple "levels" of
the stack. This commit adds a support for having multiple registered LTE
event handlers.
Note: deregister is not anymore possible with
lte_lc_register_handler(NULL),and now on, there is own function in
API for doing that: lte_lc_deregister_handler().
From that point of view, the change is not backwards compatible.
Jira: NCSDK-10078

Signed-off-by: Jani Hirsimäki <jani.hirsimaki@nordicsemi.no>