lib: location: Fixed configuration and flagging of methods
When methods were configured out, library didn't compile.
This is now fixed and it even works somehow.
There are cases, e.g., with MOSH that asserts when you have one method
compiled but you give two, or a method that's not compiled.
This will get fixed when configuration input validation has been done.
Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>