cmake: Add support for variable offset and slot size
To support the use case of multi-image MCUBoot the slot size and offset
of the signing proceedure needs to be configurable. This commit will add
these as variables to be put into the signing command.

It also shares the variable `CONFIG_UPDATEABLE_IMAGE_NUMBER` from
mcuboot to other images so that it can be used for setting different
configurations and sizes.

This commit also adds the generation of a dfu zip file with manifest for
network and application core when building for nRF53.

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>