cmake: partition manager: conditionally share b0n signed image
Only share the b0n signed image when secure boot is enabled.

Currently PM_SIGNED_APP_HEX is always shared when a domain is having
mcuboot enabled.
However, this is not the same as having secure boot enabled, so the
sharing of the non-existing signed_by_b0_app.hex file may cause other
parts of the build system to fail.

Therefore the sharing of signed_by_b0_app.hex is now safeguarded with
if(CONFIG_SECURE_BOOT)

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>