lib: lte_link_control: Notify UNKNOWN network status upon invalid CellID
Notify the application with `UNKNOWN` network registration status if the
Cell ID parsed from a `CEREG` response is `UINT32_MAX` (FFFFFFFF).

This is to fix the issue where the link controller would notify the
application with network registration statuses 1 (`REGISTERED_HOME`) and
5 (`REGISTERED_ROAMING`) accompanied with an invalid Cell ID. This would
give a false impression that the device was connected to LTE, which
would cause subsequent attempts of data transmission to fail.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>