unit_test: gps_module: Add setup and tearDown functions
The setup and tearDown functions thus added will be invoked for every
test case. They initialize the mocks and verify that all expected
functions are called the correct number of times by the unit under
test (in this case, the gps_module).

If this is not done, the test case will pass even if the the gps_module
does not call make an expected function call.

This is needed because we are using custom scripts to generate test
runners in ncs.

See https://github.com/ThrowTheSwitch/CMock/issues/202

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>