scripts: get_pulls_in_range: refine guess_pr_area()
This helper function guesses which area of zephyr development most
closely matches a pull request, given a list of commits in that PR.

Since 'Testing' was added to the list, we need to refine this function
to try to avoid returning that result for PRs that have commits with
an equal number of test case related commits as commits in some other
area 'foo'. Such PRs are likely about 'foo', not testing.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>