tests: kconfig: add option to include net core image
Currently the 'hello_world' sample is included as a child image
if you by some chance have set some random kconfigs.

This is not a clean solution as you might not want the
hello_world sample to be included when you have those
options set.

Solve this by introducing a new option which can be set if
you need to have the hello_world sample in the net core.

Ref: NCSDK-11027

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>