asset_tracker_v2: moving board specific header files to include folder
With the introduction of application configuration directory in the
boards folder, see:
https://github.com/zephyrproject-rtos/zephyr/pull/38229

then `boards/nrf9160dk_nrf9160_ns` and `boards/thingy91_nrf9160_ns`
are detected as configuration directories which causes the build to fail
because there are no project configurations in those folders.

The headers located in the boards folders have been moved to
`include/<BOARD>` instead.

Having headers in an include folder is also aligned with how headers are
organized elsewhere.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>