cmake: sign: added bin file as output from validation_data.py
The custom command invoking validation_data.py creates both a hex file
and a bin file, however only the hex file was specified as output.

Specifying the bin file as output allows other parts of the build system
to depend on the output and be correctly updated when said output
changes.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>