partition_manager: fix check of mcuboot secondary to use PM
instead of CONFIG. The size of the secondary is available
through PM_ so use this instead of the CONFIG symbol.

This way the mcuboot child image does not have to pass
the value of the CONFIG symbol to its parent image, and
statically defined secondary sizes will also be checked.

Ref: NCSDK-11237
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>