cmake: ncs package: move specification of min required toolchain
This commit moves the specification of which nRF Connect SDK Toolchain
that is minimum supported from the `NcsConfig.cmake` file and into a
dedicated `scripts/ncs-toolchain-version-minimum.txt` file.

The ensures that knowledge of minimum toolchain version required is
placed alongside other version files like: `requirements-*.txt` and
`tools-versions-*.txt`.

This should reduce the risk of developers forgetting to update the nRF
Connect SDK Toolchain version when required version of tools are updated
and thus an older toolchain can no longer be used.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>