samples: bluetooth: Stop advertising and scanning before printing logs
This sample acts as both scanner and advertiser concurrently, allowing
it to connect to another device running the same image.

This change mitigates a race condition in which the device enters two
connections when the incoming and outgoing connection is established at
the same time.

Moving the logging after the stopping of new connections decreases the
opportunity of this race to occur and seems to eliminate it in tests.

Signed-off-by: Ryan Chu <ryan.chu@nordicsemi.no>