mcuboot: hooks: Add version number 0.0.0+0 to ram flash image header
Since we don't have this information available in the application core,
we fake the data returned. So instead of it being uninitialized, we just
set everything to 0. So that it looks more intuitive for the user when
fetching the image version number.

This is mainly to be used for MCUBoot Serial Recovery as I would
expect that the `smp_svr` based sample could at some point use an custom
HCI command to fetch the image version number, build number or hash from
the running network core image.

Ref. NCSDK-11270

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>