samples: nrf9160: Initialize modem library during `SYS_INIT`
Initialize modem library during `SYS_INIT` for samples that include
the AT host library.

The AT host library is initialized during `SYS_INIT` and depends on the
AT -command and -notification libraries which in turn depend on the
modem library. Before this patch the AT host library would fail to
register an AT notification handler due to the library being
initialized prematurely.

Closes NCSDK-11273

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>