applications: asset_tracker: add multicell option
This expands the cellular positioning support to include
multicell in addition to the original single cell.

This also ensures that in single cell mode, the RSRP
is included, by adding it to modem_info.

When CONFIG_NRF_CLOUD_CELL_POS is defined, request that the
cloud report the location back to the device.

In the general data receive processing occurs, check if
it is a valid cell pos response.  If so, log the location.
If not, fall through so other handlers can take a look,
or log an error message if appropriate.

Update README.rst to add cellular positioning information,
remove a redundant section on A-GPS and P-GPS, and
document LED behavior in multi cell mode.

Jira: CIA-427

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>