scripts: generate_zip: provide option to include firmware revision
Add an optional --meta-info-file flag that can be used to pass the
Zephyr build meta file to the generate_zip script.

When the Zephyr build meta file is passed, the revisions of Zephyr and
nrf repositories will be included in the manifest file.
This is required by nRF Cloud.

Set the name field in the manifest to the value of the --name value,
if provided.

Jira: NCSDK-11510

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>
Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>