net: lib: nrf_cloud: improve non-MQTT request handling
When not using MQTT, a request for more P-GPS data is handled by the
application via the PGPS_REQUESTING event.

If the application is unable to successfully submit the request to the
cloud, it should call the new function nrf_cloud_pgpgs_request_reset()
so the P-GPS library can attempt it again later.

Regardless of transport, the function nrf_cloud_pgps_process() should
always force the state to PGPS_LOADING and set up the internal data
structures to prepare for incoming data.

Jira: NCSDK-11580

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>