[DNM] important: action needed: integration_platforms for tests and apps
Dear Codeowners,
You are addressed in this PR because you are listed as maintainers for at least one of the tests or apps in sdk-nrf/tests/ and sdk-nrf/applications.
We were successfully able to start executing CI tests in sdk-nrf in the "integration" mode for sdk-nrf/samples. Now is the time to address tests and applications. Explanation on why we need such mode is given in the comment below (copied from #4444)
Action required: Please check if tests defined in sample/testcase.yaml in your test's/app's directory contain "integration_platform" list. If not, please create a patch updating the yaml with such entry, defining on which platforms your test/app will be built in the CI. The easiest is to copy platforms from "platform_allow" entry (if it exists). Please provide such a PR before 06.10
Cheers,
Maciej

/tests/bluetooth/tester/                  @carlescufi @trond-snekvik
/tests/drivers/flash_nop_device/          @de-nordic
/tests/lib/hw_unique_key*/                @oyvindronningstad @Vge0rge
/tests/lib/modem_jwt/                     @SeppoTakalo
/tests/modules/tfm/                       @hakonfam
/tests/modules/mcuboot/external_flash/     @hakonfam @sigvartmh
/tests/subsys/profiler/                   @pdunaj @MarekPieta
/tests/subsys/zigbee/                     @tomchy @sebastiandraus
/tests/subsys/bluetooth/mesh/             @trond-snekvik

/applications/asset_tracker/              @jhn-nordic @jtguggedal @rlubos
/applications/asset_tracker_v2/           @simensrostad @jtguggedal @jhn-nordic @coderbyheart
/applications/connectivity_bridge/        @jtguggedal @nordic-auko
/applications/machine_learning/           @pdunaj
/applications/matter_weather_station/     @Damian-Nordic @kkasperczyk-no
/applications/nrf_desktop/                @pdunaj
/applications/pelion_client/              @pdunaj
/applications/serial_lte_modem/           @junqingzou @lats1980 @rlubos

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>