lib: nrf_modem: Allow to create native sockets with offloading enabled
Update west.yml to pull Zephyr changes which allow to assign priority to
the socket implementation.

With socket priorities avaialble, it's possible to ensure that offloaded
socket implementation takes preference before a naitve one, therefore
it's possible to "hijack" the socket call and disable offloading
temporarily when it's needed to create native socket.

It is now possible to create a fully native socket (with both TLS and
TCP native implemetation) or TLS only (where native TLS is used and
offlaoded TCP/IP).

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>