cmake: support for nRF Connect SDK (Ncs) CMake package
This commit introduces a real nRF Connect SDK CMake package that can be
used to lookup an nRF Connect SDK release in an out of workspace
environment.

This further allows users to match a specific nRF Connect SDK
installation when multiple releases are available in the system.

This can be done as:
> find_package(Ncs 1.7.0 EXACT REQUIRED)

The Ncs CMake package will chain load the Zephyr package into the build
system.

This means that doing:
> find_package(Zephyr)

or
> find_package(Ncs)

will have identical behavior, but with the possibility for the user to
ensure an exact version of nRF Connect SDK is used.

To use the nRF Connect SDK CMake package it must first be exported which
can be done by invoking:
> cmake -P <ncs>/nrf/share/ncs-package/cmake/ncs_export.cmake

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>