kmu: Apply PAN workaround to key_written
Fix NCSDK-11684 by adding a __DSB instruction in between the
NRF_UICR_S read and the next flash read.

This workaround and bug is described in errata for nrf91 with ID "[7]
KMU: Subsequent accesses between info_mem and main_mem of the flash
may not work properly".

The fix can be verified by applying the patch to SHA
24c60183e05341b6bbff944b45f8eb0a408710c5 and observing that this fixes
NCSDK-11684.

The fix can not be verified at the current master revision because the
PAN is dependent on the exact timing between LDR instructions and is
therefore subject to code alignment and cache misses. NCSDK-11684 was
accidentally "fixed" (more like hidden) in SHA
f72e64068f225a26a9bc8482aa914f52109edce5 just by realigning the code a
bit.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>