samples: multiprotocol_rpmsg: Increased BT_BUF_CMF_TX_COUNT
It was spotted that at the end of Device Firmware Upgrade
over BLE process appears warning related to the
HCI_ACL_FLOW_CONTROL and results with breaking further
BLE communication. Increased BT_BUF_CMD_TX_COUNT that
was introduced also for hci_rpmsg image as a workaround
for this problem.

Jira: KRKNWK-11608

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>