cloud: kconfig cloud configuration mutual exclusive option
This commit introduces the possibility of making the selection of cloud
transport service a choice.

This allows samples / applications to specify that the cloud service
options are mutual exclusive instead of individually selectable.

This is done by setting the `CLOUD_SERVICE_SELECTOR` internal variable
to `choice`.

The benefit of this is that all Kconfig settigs keeps their configurable
names regardless of whether settings are inside a group or not.
For example, `CONFIG_NRF_CLOUD_MQTT=y` is valid both in default mode
but also when `CLOUD_SERVICE_SELECTOR=choice`

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>