samples: bluetooth: llpm: Make device roles selectable over UART
This commit changes the way the master/slave role is assigned in the
LLPM sample. Instead of assigning the roles automatically, the user is
asked to select the role on each device by typing either 'm' or 's' over
UART. This fixes a race condition where a device could end up being
assigned both the master and the slave role at the same time.

Signed-off-by: Olav Thoresen <olav.thoresen@nordicsemi.no>