samples: nrf9160: download: Remove at_cmd and at_notif code
Previously, at_cmd and at_notif had to be initialized because
the nRF9160 key management needed it. This is no longer the case,
so the at_cmd and at_notif code is removed.
The changes are needed as preparation for changes in
LTE link controller to use nrf_modem_at as AT interface.

In addition, nrf_modem is set to use sys init and lte_link_control
is set to not auto-connect. In this sample, there is no need for
nrf_modem to be initialized by the application, so this is done
to slightly reduce the complexity of the sample.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>