tests: event_manager: Fix sys_reboot misuse
`sys_reboot()` function is expected to be non-return function (marked
now explicitly with FUNC_NORETURN attribute). As the OOM event_manger
test case reimplements this function as returning, the compiler
complains with a warning and the code crashed when executing test.

Fix this by implementing the function as truly non-return. Since the
test case is consdered successful when `sys_reboot()` is called, end it
explicitly from there, releasing resources beforehand.

Finally, rework the rest of the test, to be executed directly from the
test thread instead of system workqueue via event manager. The reason
for this, is that after non-returning `sys_reboot()` is called from the
workqueue context, it will remain blocked. When calling `sys_reboot()`
from the test thread, the test suite takes care of aborting the test
thread and starting fresh for the next test case.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>