net: lib: aws_iot: Detect and handle MQTT ping failure
If an error is returned from mqtt_ping(), the connection
has likely closed due to a NAT timeout by the carrier or a
keep-alive timeout by the broker.

Before calling mqtt_ping(), check client.unacked_ping; if
non-zero, the previous ping was never acknowledged, which
indicates the connection was closed for similar reasons.

In aws_iot_cloud_poll(), check for and handle this error.

Jira: CIA-439, CIA-440

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>