nfct: Drop support for compiling without NFCT defines present
There is no reason for FICR_NFC_TAGHEADER0_MFGID_Msk to not be
present. And if, for whatever reason it is missing we should fail with
a compilation error instead of silently using, presumably, invalid
values.

Also leave a TODO for the next developer that adds support for TF-M
with NFCT.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>