lib: lte_link_control: Move cell ID check under roaming/home reg status
Check if cell ID is invalid when registered to home or roaming network.
If invalid, the `LTE_LC_NW_REG_UNKNOWN` network registration status
is propagated. This is handled internally for modem firmware versions
1.3.0 +.

Under certain circumstances, an invalid cell ID is expected. The check
was moved to ensure that the correct registration status is notified
in those cases.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>