samples: nrf9160: gnss: add nRF Cloud A-GPS and P-GPS support
Added support for nRF Cloud A-GPS and P-GPS using the REST API.

When assistance is used, LTE now remains connected all the time
by default. Both eDRX and PSM are requested from the network.
It's possible to enable the old behavior using a new Kconfig
option so that LTE is only activated when needed to fetch A-GPS
data.

LTE link control is used to configure modem also when assistance
is not used.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>