Bluetooth: Mesh: Do sampling with the fastest cadence
When the publication period is significantly bigger than the fast
cadence interval, it may take the whole publication period to before
server notices that the sensor change is within the fast cadence
interval.

This change makes the sensor server run publication timer with the
fastest cadence so that in worst case the change in a sensor
will be recorded at the next fast cadence period.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>