Bluetooth: Mesh: Add a note about end of transition publication
According to MshPRFv1.0.1, section 3.7.6.1.2, when publishing is enabled
for a model, unless specified otherwise, by a higher layer
specification, a corresponding status message shall be published
immediately after the state transition ends.

This behaviour is not implemented in the stack. The models only publish
at the start of transition. An application is responsible for publishing
a status message at the end of transition.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>