samples: bluetooth: enable tinycrypt for le sc pairing
The Softdevice controller does not support ECDH HCI commands that are
required by the LE Secure Connections. To workaround it, the
BT_TINYCRYPT_ECC configuration is used to emulate ECDH HCI commands
at the Bluetooth Host layer. This Kconfig option is set by default if
the BT_CTLR is also present. This is not the case for the nRF5340
application image. That's why the BT_TINYCRYPT_ECC config needs to be
explicitly set in the prj.conf file.

Ref: NCSDK-12190

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>