mcuboot: imply BUILD_OUTPUT_META when enabling SIGN_IMAGES
imply'ing BUILD_OUTPUT_META will help to ensure that zephyr.meta file
is created as part of the build and thus include firmware revisions
in generated dfu zip package.

imply is used to allow end-user to disable revisions in the dfu
firmware package if they have reasons to not want this behavior.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>