Bluetooth: Mesh: Don't use sensor cadence state if it is not configured
This reflects changes in E15886 which states in that values of sensor
cadence state shall be ignored if it is not configured.

Section 4.3.1.2.4.2:

If the Sensor Cadence state has not been configured, the Sensor Cadence
state values shall be ignored, and any change in the Sensor Data state
that is greater than an implementation-specific minimum amount shall
trigger immediate publication of the Sensor Status message. This
implementation-specific threshold should be chosen to prevent frequent
small changes in the raw sensor data resulting in publication of an
excessive amount of data.

Section 4.3.1.2.4.3:

If the Sensor Cadence state has been configured, the cadence of the
periodic publication shall be modified as follows...

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>