fatal_error: Only default to RESET_ON_FATAL_ERROR for ncs samples
Don't default to rebooting on fatal errors.

Rebooting causes loss of information and is a pain while debugging. It
can also have unintended side-effects in real applications.

I can see the use in certain types of applications, but this should
not be default behaviour for all NCS builds.

This changes makes it the default only when NCS_SAMPLES_DEFAULTS is
set as it can be useful for demonstration purposes.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>