cmake: modules: add referenced memfault-firmware-sdk Kconfig symbols
The `memfault-firmware-sdk` project is not a crucial project in the
nRF Connect SDK, and downstream users should be able to remove this
project from their downstream manifest file.

But the nRF Connect SDK Kconfig is referencing upstream Kconfig symbols
which results in errors, even when building hello_world`, for example:
> warning: <choice MEMFAULT_HTTP_PERIODIC_UPLOAD_CONTEXT>
>    (defined at .../Kconfig:174) defined with type unknown
>
> warning: <choice MEMFAULT_HTTP_PERIODIC_UPLOAD_CONTEXT>
>    (defined at .../Kconfig:174) defined without a prompt
>
> ...
>
> error: Aborting due to Kconfig warnings

by adding the referenced upstream symbols and safeguard them inside
`if 0` then the symbols is known, which avoids the error.
And because the safeguard, the symbols are not configurable.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>