asset_tracker: remove nmos code
Although ui_nmos_init() is called to initialize the transistor gate
pins, nothing else in the nmos.c translation unit is ever called.

Since the drain and source of the transistors are only available on
external connectors, and we aren't doing anything with the gates, this
entire module is essentially dead code. Just remove it in order to
avoid having to convert its Kconfig symbols to devicetree equivalents.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>