asset_tracker: remove CONFIG_UI_LED_PWM_FREQUENCY
This isn't being set anywhere in-tree, so just remove it to keep an
upcoming conversion to devicetree simple. We can always add it back as
a devicetree property if it is really being used somewhere, and
developers hacking around can change a new C macro which is used for
the previous Kconfig default.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>