samples: nrf9160: lwm2m_carrier: Add AT+CNUM failure handling
Handle an error cause where SIM does not have MSISDN.

This error happens on a new SIM which have not been registered on
the network yet. Handle this by waiting for 3 minutes to let the
registration happen, and then reboot to initialize the lwm2m_carrier
library again.

Retry up to 3 times to avoid an infinite reboot loop in case of
other errors.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>