lib: bin: lwm2m_carrier: Add AT+CNUM failure handling
Handle a situation when SIM does not have subscriber MSISDN.

This happens on new a SIM which have not been registered on the
network yet. Handle this by retry every 10 seconds for up to
10 minutes to let the registration happen.

Add check for AT error code in lwm2m_os_at_cmd_write() when
application has enabled AT+CMEE=1. A valid CME error code 0
(Phone failure) should not be reported as success.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>