scripts: tweak west ncs-loot warning output
The ncs_west_helpers.py script prints a warning when a project
has a commit with a bad or missing sauce tag. This appears
in a confusing place in the output though, because it is printed
before the banner naming the project and its path.

Due to the way the code is structured, moving this warning to after
the banner is not possible to do easily, but make the situation
less confusing by printing the project name.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>