net: lib: Increase MQTT payload buffer size in Azure IoT hub and AWS IoT
Increase the default MQTT payload buffer size in
Azure IoT hub and AWS IoT to 2304 bytes.

This is to address a common issue where users sends messages to the
device exceeding the maximum payload buffer size, which triggers an
error. The new default MQTT buffer size aligns more with the maximum
size of the internal TLS buffer in the modem.

Closes CIA-337

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>