CODEOWNERS: Replace ioannisg with SebastianBoe
ioannisg is unavailable so we replace all review requests with myself.

I have only taken over the TF-M responsiblity from ioannis, so this is
still wrong. But it is better to have the wrong reviewer than an
unresponsive reviewer.

The correct mapping will be cleaned up in january. Until then I will
re-assign anything that comes in that is not TF-M related.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>