caf: sensor_manager: Fix incorrect error recovery
Sensor could incorrectly recover from ERROR state
when power_down and wake_up events are processed.

Checking state of a sensor must be done in non preemptive
context. We can not assure the events are handled in
cooperative context so we must lock the scheduler when
checking state of each sensor.

Jira: NCSDK-12783

Signed-off-by: Emil Obalski <emil.obalski@nordicsemi.no>