tests: lib: at_cmd_parser: Add extra checks for string lengths
Static analysis tool detected that the buffer length to compare might
exceed the provided string length, which is not correct as the
at_params_string_get() should overwrite the tmpbuf_len value with the
actual string length read.

Make sure that the result string length matches the expected one before
doing memcmp to prevent such reports in the future.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>