kconfig: mcuboot: fix 16kB minimal builds
When external flash and minimal build is enabled the MCUboot partition
should be 0x3e00.

Due to locking granularity, it is important to be able to fit two
MCUboot images within 32kB when B0 is enabled (implied by
BOOT_NRF_EXTERNAL_CRYPTO being set). Otherwise more memory will
be needed to support the update mechanism of B0 (32kB extra on
nrf53 and 64kb extra on nrf91).

Ref: NCSDK-12199
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>