lib: fota_download: Button re-enable after err in download_with_offset
After disconnecting from the net in the middle of a download,
fota_download have the ability to restart from the point it left off
when dfu is re-started. In this case, the download_with_offset()
function will be called.

If an error happens in the download_with_offset() function,
the dfu is stopped and the button used in the http_update samples for
starting the dfu will not have been re-enabled. This makes it
impossible to restart dfu without resetting the board, in case of
an error in download_with_offset().

NRF91-1287

fota_download will now send error event to application so application
can re-enable the button.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>