application: asset_tracker_v2: Enable `doreply` flag in cell pos request
Enable `doreply` flag in cell position requests and filter (ignore) the
corresponding responses carrying the approximate location of the
device.

Due to ongoing changes in nRF Cloud regarding visualization of location
based on cellular measurements, it might be necessary to request
the response to cellular position request. If this is not needed anymore
this commit will be reverted in the future.

In general, the application does not care about getting the location
back from cellular position requests. However, this commit is
needed to ensure that the cellular location of the application on
tag v1.8.0 will be visualized in the nRF Cloud web UI, regardless.

Fixes CIA-477

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>