lib: bin: lwm2m_carrier: Non-blocking connect to LTE network
Use non-blocking connect to LTE network to avoid an error condition
in lwm2m_carrier_init() and lwm2m_carrier_run() when a device boot
in an area without LTE coverage or bad reception conditions.

The event handler is not in use by the LwM2M carrier library and
can be utilized by the application if needed.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>