net: lib: lwm2m_client_utils: FOTA image filtering for modem
Modem firmware can be provided either as full image or delta image.
We can distinguish between the images when we start the download, but
not before. As FOTA-library doesn't currently support expecting multiple
images, we start expecting modem delta image. If we notice that it is
actually a full image, we restart the fota download expecting a full
image.

Signed-off-by: Jarno Lamsa <jarno.lamsa@nordicsemi.no>