samples: nrf9160: http_update: graceful shutdown of modem fw
When updating firmware with http_update sample(s), one or more resets
must happen in order to apply the new image. These resets are causing
ungraceful disconnects/reconnects from the network, and the modem
will after a number of reconnects like these, refuse all
subsequent reconnection attempts for a set amount of time.

To avoid this, the sample(s) now gracefully disconnect from the
network by setting the modem in the power off mode when the
firmware image download is complete.

Signed-off-by: Andreas Moltumyr <andreas.moltumyr@nordicsemi.no>