[nrf mergeup] manifest: Upmerge to jan. 5 Zephyr SHA
Upmerge from 2.7.0 to the Zephyr main branch on 5. january.

Upmerge mcuboot to 399720d1cabd26c4356445d351f263b31e942961,
which was the mcuboot revision used by Zephyr on January 5.

Add the testing framework Zephyr module fff to the name-allowlist.

This could be required to run certain tests in the future.

Add the scientific/math Zephyr module zscilib to the name-allowlist.

This could be a useful replacement for gsl / Lis / CMSIS-DSP.

Add the two TF-M Zephyr modules psa-arch-tests and tf-m-tests. These
used to be copied into the tfm module. But now exist as their own
modules.

These are required for integration testing of TF-M.

tfm used to be a repository that had the contents of three
repositories copied into it. This tfm repository no longer exists. In
it's place are the three repositories that used to be copied into tfm.

trusted-firmware-m
psa-arch-tests
tf-m-tests

Use an NCS fork of mbedtls instead of upstream mbedtls.

trusted-firmware-m: upmerge jan. 5

Upmerge from 1.3.99 to 1.5.0 (+ downstream patches).

The 1.3.99 revision contained 3 repo's, including
trusted-firmware-m. This new revision is the trusted-firmware-m
repo. So the repo looks completely different.

History has not been retained as it is in practice a new repository.

The sauce tag commits were re-applied to this new repository.

Update sdk-mbedtls from 2.26.0 to 3.0.0.

3.0.0 is the same as zephyr's mbedtls except that zephyr has some
extra patches.

This updates ncs/mbedtls, as opposed to
ncs/modules/crypto/mbedtls.

Eventually, we will remove this so there are not so many copies of
mbedtls.

This update fixes some integration issues with nrf_security.

Remove the mcumgr module which is now copied into Zephyr instead.

Bumped nrfxlib to a revision that is compatible with new Zephyr
revision.

Changed sdk-connectedhomeip fork revision to support mbedTLS 3.0.0
version API in Matter.

Updating to nrf_security with mbed TLS 3.0.0

Note: Single commit to keep bisectability. There are configuration
changes in nrfxlib/zephyr that is dependent

-Add west.yml checkout of new nrflib (PR) - With mbed TLS 3.0.0 support
-Add west.yml checkout of new zephyr (PR) - With PSA configs
-Changed configurations for PSA crypto example
-Removed invalid backends for tests/crypto
-Updated tests/crypto to work with external RNG
-Modifications for test-examples for mbed TLS 3.0.0 support
-Added defconfigs for PSA for TF-M usage

ref: NCSDK-11689

Update sdk-homekit to be compatible with the upmerge.

Update sdk-find-my to be compatible with the upmerge

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>