upmerge: pm: Port to new pm_device_ API
There are some breaking changes to the power management API.

pm/device.h needs to be included. Presumably this came from device.h
before.

pm_device_state_set no longer exists. Nor does there exist a
one-to-one replacement for it. But pm_device_action_run is close.

Instead of requesting states, one requests actions, and hopes that the
desired state is achieved.

There is a new enum for describing the actions. This means that one
can no longer do.

function(target_state)
change_state(target_state)
endfunction()

Instead one must translate from the target state to some action, and
then do the action. Or drop support for requesting target states.

The device.h API does not support such a translation, so we would need
to translate ourselves.

In some cases we have been checking the current state, and then
comparing it to the target state before requesting the target
state. This is awkward with the new API.

In these cases we are now instead skipping the check of current state,
and just doing the action that should give us the desired state. If
the state was already set correctly we get a -EALREADY error code
which we suppress. This is nearly functionally the same as checking
the current state first.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>