drivers: timer: Port to new SYSTEM_CLOCK_DISABLE API
CONFIG_SYSTEM_CLOCK_DISABLE no longer exists.

It used to add a function to the kernel. In it's place is
CONFIG_SYSTEM_TIMER_HAS_DISABLE_SUPPORT.

This new option works slightly differently. It also adds the function
to the kernel, but instead of the user enabling it, it should be
selected by any provider of this kernel function.

The Kconfig option was changed in this commit:
https://github.com/zephyrproject-rtos/zephyr/commit/
78dc8ce33828b1d995d520751fe761ade6103155

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>