doc: mcuboot: do not copy all .md files
We can't copy anymore all *.md files, since some of them reference now
documents outside of the docs/ folder. We have been lucky that none of
the required documents do this, however, this may happen in the future.
This shows that the MCUboot docset approach is not sustainable in the
long term, so a better way needs to be found.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>