applications: asset_tracker_v2: Explicitly disable PSM if not enabled
If the user sets CONFIG_MODEM_AUTO_REQUEST_POWER_SAVING_FEATURES=n,
it's reasonable to expect that PSM would not be requested by the
device.
However, the nRF9160 modem stores PSM request information to NVM,
so PSM must explicitly be disabled before the modem is put into
normal mode to be sure that PSM is not requested by the device.

This patch adds such logic to the modem module.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>