lib: at_monitor: Increase maximum allowed AT monitor heap size
In some cases, several AT notifications can come in rapid succession
and require more than the previously allowed maximum heap size.
The issue is more easily triggered if one or more of the notifications
are large, like for instance an %NCELLMEAS notification with many
cells.

This patch increases the upper limit of the heap size range to let
applications set a higher value if required for their use-case.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>