applications: asset_tracker_v2: Add new Kconfigs to convert RSRP/RSRQ
Depending on the integration, RSRP and RSRQ values need to be provided
unconverted or in dB/dBm in the various sample events sent out by
the modem module. This is due to the application using different
encoding libraries depending on the data that is sent and
the configured cloud configuration.

This patch deprecated `CONFIG_MODEM_CONVERT_RSRP_AND_RSPQ_TO_DB`

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>