samples: Bluetooth: df: Fix stack overflow issue
Host receive thread has small default stack size.
If periodic advertising is enabled max amount of advertising
data is set to 1650 bytes. Attempt to convert data from binany
to string requires large array. Allocation of such array on
stack caused stack overflow.

To limit the scope of the array and avoid stack overflow
the array is changed to be static local variable in a function.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>