tfm: Update flash_layout for Non-Volatile Counters changes in TF-M
Update flash_layout to define the required definitions for
OTP / Non-Volatile Counters area.
Split the TFM extra partition into three, so that the partition is
properly sized based on which features are currently enabled.

TFM_HAL_PS_PROGRAM_UNIT and TFM_HAL_PS_SECTORS_PER_BLOCK need to be
defined for ITS when PS is disabled.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>