doc: glossary: fix "board"
This might be a bit nitpicky, but this is a foundational concept and
its current glossary definition has some problems:

- it is not a data structure
- it is much more general than development kits
- a "board" doesn't inherently specify secure/non-secure variants

Strictly speaking, a "board" in the Zephyr context is a named hardware
target that has a software support package the build system can use to
build Zephyr applications that run on the target. The support package
itself is files in a directory on the file system, which is not really
a "data structure" in the computer science sense:

https://en.wikipedia.org/wiki/Data_structure

Commit 65c9bf91aca5a27 ("doc: Glossary Update") changed "board" to
"development kit" in the definition, maybe to avoid a circular
definition of "board" in terms of "board". But a customer's own
hardware also must have a Zephyr board definition, and this is not a
development kit. Limiting the definition to development kits is
therefore not correct. "Hardware target" is more accurate while still
avoiding a circular definition.

The non-secure board variants whose names end in "_ns" are "just" a
set of conventions tacked onto the board concept. There's nothing
inherent in a board that specifies whether a target is secure or
non-secure. Removing this is more accurate and may avoid confusion.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>