multi_image: partition_manager: fix bug in bootlader/pm.yml
MCUboot child image kconfig options are not available when preprocesing
this file. Hence, checking if CONFIG_PARTITION_SIZE_MCUBOOT_PAD = 0
will always be true.

Add a Kconfig option which is used to pass information to non-mcuboot
child images that mcuboot is included in the build and use this
instead.

This places s1_image to an incorrect location both when mcuboot is
included and not.

Ref: NCSDK-13188

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>