scripts: unity: Fix function definition regex bug
The whitespace fix to this regex last week triggered another bug in
libmodem. They have a typedef that the regex parsed as a function
definition because it accepted too many character in the argument
list of a function. This fix updates the regex to only allow
characters you would expect in an argument list of a function def.

Characters expected in the argument list of a function def:
\w A-Z a-z 0-9 and _
\s All whitespace
,  Argument separator
*  Pointers
\. Variadic argument
\[ Array start
\] Array end

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>