tests: unity: Separate example test and wrap tests
The python scripts to prefix functions with __wrap was tested using the
example_test. This test is mean as an example of how to do unit
testing. Moved the existing wrap tests and some new ones to a new test
suite to avoid adding more irrelevant tests to the example test.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>