modules: tfm: Fix TF-M include folder location
Fix the TF-M include folder to point to the correct destination.
ZEPHYR_TRUSTED_FIRMWARE_M_MODULE_DIR has been moved but now nrf_security
adds the correct path to zephyrs includes.

Remove tfm/ path for includes. This is not needed as include/tfm is
also added to the include paths. In addition tfm_ioctl_api.h wouldn't
be in a tfm/ folder when using upstream TF-M, so this would cause
confusion with vanilla zephyr builds.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>