doc: update DK Button and LEDs
Use "buttons" and "leds" consistently in document titles and when
naming the library.

Mention the main Kconfig option used to enable the library to make it
easier for people just getting started.

Add a section on how to adapt the library to custom boards, which may
be useful to customers who are prototyping on a DK and wanting to
use the same button/LED interface when debugging custom hardware.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>