applications: asset_tracker_v2: improve P-GPS support
Add missing handling of preemptive updates after injecting data.

Add missing request for prediction notifications when A-GPS
download is in progress.

Add missing request for prediction notifications when A-GPS
download is processed correctly.

Make A-GPS work along with P-GPS if both enabled.  A-GPS is
be used for everything but ephemeris data, while P-GPS is used
for only that.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>