applications: nrf_desktop: Default size of idle thread stack
The latest changes in Zephyr require us to increase the idle thread
stack size a bit if it was set to 128 bytes.
I have checked and it seems we have this setting inconsistent
between boards and even one board. We also have space in RAM.
Let's then use the default stack proposed by Zephyr.

Signed-off-by: Pawel Dunaj <pawel.dunaj@nordicsemi.no>