lib: nrf_modem: Revert modem trace change from #6471
It was assumed that the modem trace was always enabled at boot, but the
modem trace configuration is persisted in the modem periodically and
when using AT+CFUN=0. This caused a regression when the modem did not
already have the modem trace enabled in the persisted configuration.
This commit reverts the changes in lib/lte_link_control/lte_lc.c from
should always be enabled when CONFIG_NRF_MODEM_LIB_TRACE_ENABLED is
turned on.

Fixes NCSDK-13500

Co-authored-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>
Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>