Bluetooth: Mesh: Make regulator to not interfere with controller's FSM
Current implementation may produce the light's glitches, because the
output of the FSM and the regulator's output interfere when FSM
transitions to another state.

When the regulator is enabled, only the maximum of 2 states, the
Lightness Out and Regulator Output, should be published (section
6.2.3.6). If no sensor reports have been provided yet, the regulator
will drive the lightess to the maximum value. To avoid this, the
controller should publish the Lightness Out state until any sensor
report with an ambient light level is provided.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>