lib: lte_lc: Add checks to connect_lte() to avoid hanging
If an application calls lte_lc_connect() or lte_lc_init_and_connect()
when the device is already registered with an LTE network, it would
previously hang until timing out, waiting for the link semaphore that
was already given.

A similar issue could occur if two threads called one of the APIs
at the same time. The semaphore would only be given to one of them,
and the other thread would hang.

This patch introduces two mechanisms to avoid this:
- Check network registration status before doing device configuration
  and setting the modem's functional mode.
- A flag to indicate that a connection establishment is already in
  progress.

Fixes CIA-542

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>