lib: lte_lc: Add cell ID check in lte_lc_nw_reg_status_get()
A check for valid cell ID was missing in lte_lc_nw_reg_status_get().
The check is needed because older modem firmwares may
report a combination of home or roaming registration status
even if no valid cell ID is present. In that case, the registration
status should be set to 'unknown'.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>