samples: matter: change maximum light brightness level
Matter samples assumed that the maximum brightness of the
simulated light bulb is 255 instead of 254, as used by the
level control cluster. As a result, the corresponding LED
would never stay continuously on. The difference is too
negligible to be noticed by a human eye, but it enough to
cause intermittent automated test failures.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>