doc: extensions: initial version of manifest_revisions_table
This extension allows to render a table containing the revisions of the
projects present in a manifest file. To do that, it introduces a new
directive: ``manifest-revisions-table``. It can be used in the code as:

    .. manifest-revisions-table::
        :show-first: project1, project2

where the ``:show-first:`` option can be used to specify which projects
should be shown first in the table.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>