doc: hw_unique_key: Remove conditionals in header file
Remove the conditionals in the API header file for HW unique key
feature. There is no valid reason to compile out the difiniton of the
functions when the feature is not enabled.
With the definitions compiled out they are excluded from the
documentation unless they documentation build "enables" the features.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>