tests: lib: nrf_modem_lib: test socket offloading
This commit adds unit tests for the socket offloading
layer, according to POSIX.1-2017.

This commit aligns the socket offloading layer to comply
with the unit tests:

* Fixes inconsistent behavior in `nrf91_socket_offload_accept`
in case `nrf_accept` returns a negative number that is not -1;
the new code now will call `nrf_close` on the returned value
only if it's a valid file descriptor number.

* Add NULL verification in `setsockopt`.

* Solves segmentation fault in `recvfrom` when providing
`from` not NULL and `fromlen` NULL.

* Now `getaddrinfo` returns `EAFNOSUPPORT` instead of
`EPROTONOSUPPORT` when family is not supported.

* Now `bind` returns `EAFNOSUPPORT` instead of
`ENOTSUP` when family is not supported.

* Calling `connect` with raw parameters (non-IP) will
not override the error codes set by the modem library.

* Now `sendto` returns `EAFNOSUPPORT` instead of
`ENOTSUP` when family is not supported.

* In `recvfrom` add check for expected `sock_len`
output value.

Signed-off-by: Mirko Covizzi <mirko.covizzi@nordicsemi.no>