cmake: use host system list delimeter when updating environment PATH
Jira: NCSDK-13926

When using NCS Toolchain, the environment path is prepended with the
path to the NCS toolchain to ensure that scripts (Python/CMake) and west
includes the NCS Toolchain path when invoked.

Initially NCS Toolchain was only available on Windows, but as it has
also been made available on macOS, update the code so that it uses the
correct environment list delimenter based on the host system.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>