partition_manager: Fix non-secure and secure storage alignment
Fix non-secure and secure storage alignment.
Non-secure storage was placed after the secure storage, in this
configuration they would have to be aligned on the SPU flash region
size. Move it before the secure storage to avoid this.

Secure storage was aligned using the wrong configuraion.
This was not an actual bug since they are equal, but could be a problem
in the future.

NCSDK-12968

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>