scripts: sbom: Add "west ncs-sbom" tool.
The "west ncs-sbom" tool creates a Software Bill of Materials.
It uses a built application or the specific files as an input,
detects their licenses, and creates an output report that can
be in HTML or SPDX format.

It is significantly different than the Zephyr's "west spdx"
tool, so it was implemented from the scratch. It has a different
method of getting files from the application that gets a more
precise list of files. It has also additional methods for
verifying that the list is correct. It adds more license
detection methods including an external detector from
"scancode-toolkit". The output (both HTML and SPDX) is more
readable.

Signed-off-by: Dominik Kilian <Dominik.Kilian@nordicsemi.no>
Signed-off-by: Marcin Jeliński <marcin.jelinski@nordicsemi.no>