samples: nrf9160: modem_shell: Disable UART1 by writing to registers
Previosuly the uart1 was enabled in device tree overlay mainly for the
device_get_binding to work on it so that it may be disabled. This is not
needed because UART1 is not managed by the zephyr UART driver when modem
traces as enabled. It is rather managed by the nrfx UART driver. Hence
the device tree overlay is removed.
The UART1 is now disabled (placed in low power state) by writing to its
registers as mentioned in the "Low Power" chapter in
https://infocenter.nordicsemi.com/topic/ps_nrf9160/uarte.html .

Tested by measuring power consumption with and without modem traces
enabled. It is seen that when the UARTs are off, the average current
consumption reduces to 3.5 uA.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>