applications: matter_weather_station: fix stack overflow
The default 15.4 RX thread stack size is not sufficient
for the weather station, probably because it uses FPU
which increases the stack guard size.

Also, it turns out that 4k buffer is too small for the
amount of logs produced by the application and some
messages are dropped. Double the size, by the way.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>