samples: matter: disable RTC timer in MCUBoot
The default value of CONFIG_NRF_RTC_TIMER_USER_CHAN_COUNT
for nRF52 SOCs has been changed from 0 to 3, but it makes
MCUBoot get stuck on erasing flash pages. In fact, the RTC
timer doesn't seem to be used by MCUBoot in the case of
Matter samples, so simply disable it to avoid similar issues
in the future.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>