net: lib: download_client: use socket timeouts for coap block transfer
The download client did not implement any timeout for recv as well as
send to the socket as well it did not use a proper retry mechanism for
resending requests (for a new block).

This patch implements poll mechanism for recv (POLLIN) and send
(POLLOUT) and uses coap back-off mechanism for limiting the number of
requests that are sent
(CONFIG_DOWNLOAD_CLIENT_COAP_MAX_RETRANSMIT_REQUEST_COUNT).

Signed-off-by: Andreas Chmielewski <andreas.chmielewski@grandcentrix.net>