applications: asset_tracker_v2: Add support for modem domain events
Add support for modem domain events via the Link controller (%MDMEV).
Modem domain events can be used to notify the application of specific
events.

One of the events that the application now receives is the
RESET LOOP event. This event triggers if the device has tried to
connect to the LTE network to frequently which will restrict
subsequent connection attemps for 30 minutes.

This loop is very typical to trigger during development and can be
very tricky to debug. Therefore we now print a log warning
in case this happens.

Fixes CIA-565

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>