lib: location: support combining A-GPS and P-GPS
When both A-GPS and P-GPS are enabled, pass the
ephemerides assistance request to P-GPS, and not that
or almanac assistance to A-GPS.

In REST mode, if both A-GPS and P-GPS are enabled,
invoke the P-GPS notify work function once A-GPS
data has been received and processed.

Utilize the current prediction if included with the
PGPS_EVT_READY event.

Rely on P-GPS library defaults to enable using a dedicated
P-GPS partition instead of requiring reuse of the
MCUboot secondary partition.

Call nrf_cloud_pgps_notify_prediction() after processing
P-GPS response, so prediction gets injected when available.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>