samples: nrf9160: modem_shell: integrate with P-GPS partition
If the P-GPS partition in flash is enabled, there is no
need to pass the PM_MCUBOOT_SECONDARY_ADDRESS or _SIZE
to the init function.

Start using P-GPS partition so MCUboot is not required.

Once P-GPS response has been processed, also call
nrf_cloud_pgps_notify_prediction() so that the current
prediction actually gets injected to the GNSS unit.

If a prediction is ready right after P-GPS initialization,
go ahead and inject it.

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>