net: lib: lwm2m_client_utils:lwm2m_utils: Updating bootstrap keys
This change allows bootstrap server to write new set
of bootstrap credentials for us.

When we receive new bootstrap credentials, we should
write those to modem at the same time when we update
server keys as we don't store any PSK keys to flash.

I also splitted the init() function to two separate ones,
so it is now obvious what part is just initializing the
default object instance (ID = 0) that is created when
security object is initialized. When we have loaded
new bootstrap data from flash, we are now deleting
the default object 0 as it is now considered
outdated.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>