applications: asset_tracker_v2: Fix unit test failures on native_posix
The unit tests for debug_module and gnss_module were failing because the
mock for module_start() was checking all parameters in the input
structure. Some of those parameters can be random values and they may
make the test fail intermittently. This is now fixed by using a stub
that only checks the fields in the struct that are relevant for the
test.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>