lib: nrf_modem: Add experimental thread based trace processing
Add experimental kconfig CONFIG_NRF_MODEM_LIB_TRACE_THREAD_PROCESSING.
When disabled, modem traces will be processed synchronously like before.
When enabled, modem traces will be processed in a thread. This is to
reduce the time spent in interrupt context before passing the modem
trace data to time-consuming transports.

Disable zephyr driver for UART1 for thingy91 as done for the 9160dk.
This is to prevent the zephyr driver initializing UART1 and ending up
creating a conflicting interrupt handler for UART1.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>
Co-Authored-By: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>