doc: improve user guidance
The Getting started documentation for nRF9160 DK and
the Nordic Thingy:91 does not do a perfect job guiding the user
where they should go after completing that documentation.

This adds explicit guidance to both the start and end of
the Getting started with [product] pages. Additionally, a hint
to familiarize with the nRF Connect SDK in general is added to
the Developing with [product] pages.

Signed-off-by: Wille Backman <wille.backman@nordicsemi.no>