applications: asset_tracker_v2: Fix possible buffer overflow
When adding custom fields to the nRF Cloud publish, if the data types
names of the custom fields are longer than the buffer, it will
crash the thread with a buffer overflow.

This PR fixes that by using strncat instead of strcat, and
keeping track of how full the buffer is, logging only a partial
list of the data types requested if it would overflow to
log all of them.

Signed-off-by: Mariano Goluboff <mariano.goluboff@nordicsemi.no>