requirements-extra: don't fix pygit2, zcbor versions
pygit2 was fixed to an exact version in b837aa41d9 ("requirements: Fix
pygit2 version").

zcbor has been fixed at an exact version since it was introduced into
this file.

I continue to maintain that the requirements.txt files are not right
place for fixed versions like this, and since I am responsible for
this particular one, I'm going to change '==' to '>=' instead in the
requirements-extra.txt file.

This is because I install the latest pygit2 on my machine without
building a wheel for it with:

  pip install --only-binary=:all: pygit2

Then I install the rest of requirements-extra.txt with:

  pip install -r requirements-extra.txt

Since the first pip command will pull in the latest pygit2, not
exactly 1.6.1, the second command tries to build a wheel, which fails
for me. I believe that we need to allow developers to keep testing the
latest versions. Otherwise we will surely fall behind as upstreams
evolve. Early adopters like me should be the first line of breakage
detection where things like this are concerned. Restoring more
flexible version specifiers in requirements-extra.txt un-breaks
workflows like mine.

Since CI does need fixed versions to avoid breakage in python
dependencies, add an exact pygit2 version to requirements-fixed.txt.
The fixed zcbor entry is already in that file.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>