application: serial_lte_modem: Add multiple socket support
Add support for opening multiple simultaneous sockets. This adds
an array of socket structures to support them simultaneously.
In order to maintain backwards compatibility, this changes the
sock variable to a pointer. That points to the current socket
in the new array of sockets. This way, all the commands to
operate on a socket (like XSEND or XRECV) stay the same.

XSOCKET and XSSOCKET change to return the handle for the
socket (which is the index in the array plus one). And they
both get a new command (3) to switch the currently selected
socket.

Signed-off-by: Mariano Goluboff <mariano.goluboff@nordicsemi.no>