samples: modem_shell: Disable modem traces before disabling UART1
- When UART is used as modem traces transport medium, the modem traces
need to be disabled before disabling UART. The modem traces need to be
reenabled after reenabling UART. This is done now.
- The enabling and disabling of modem traces cannot be done from an
interrupt context because they involve sending AT commands using the
nrf_modem_at_printf() which cannot be called from interrupt context.
Hence this is done from a work queue handler.
- Also removed the logic to stop RX when disabling UART1 because the
UART1 is not used for RX. The STARTTX is also not issued anymore when
enabling UART1 to avoid spurious transmissions.

Fixes MOSH-251.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>
Co-Authored-By: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>