doc: scripts: Software Maturity Report Generation
Added a script that scans the twister build output and .config
files, and generates a software maturity level report. The
script can be run on a local twister build, but is intended to
be run by the CI and publish the results to Azure.

Also added an initial feature-file containing the desired
features and their respective Kconfig rules.

Signed-off-by: Gaute Svanes Lunde <gaute.lunde@nordicsemi.no>