applications: nrf_desktop: wheel: add support for pinctrl
Even though the QDEC pins are controlled by the QDEC driver, the wheel
code in nrf_desktop configures QDEC pins on its own. In a first place,
it should be the QDEC driver that exposes such functionality (note that
application code is using QDEC device's DT properties). The way GPIOs
are used needs to be improved as well: using gpio cells and not relying
solely on pin information (gpio_dt_spec).

Consider this patch as a workaround until a better solution is
implemented.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>