lib: nrf_modem: change log module to `nrf_modem` in `nrf_modem_os.c`
These prints are mostly from libmodem, some are from the
integration layer. Ideally we should have a log module for the
integration layer (nrf_modem_lib) and a log module for the library
(nrf_modem), but the _DEBUG_ALLOC entries prevent us from doing it
cleanly.

Rename for now, think about removing _DEBUG_ALLOC in the future,
and splitting the remaining diagnostic in a separate file.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>