lib: at_host: allow custom commands prefix
This commit adds the possibility to define a custom
at commands prefix (AT# for example) which will be
passed to the application instead of being sent to
the modem.

This can be useful for example if the nRF9160 is
connected to another core with a UART bus. It allows
sending 'normal' AT commands that will be directly
exectuted by the modem and custom-made AT commands that
will be executed by the main application.

Signed-off-by: Giuliano Franchetto <giuliano.franchetto@intellinium.com>