tfm: Configure storage partitions as non-secure with TF-M enabled
Make sure that all storage partitions used by the application is
configured as non-secure by TF-M.
Since TF-M already had a non-secure partition that was used for the
settings partition this functionality is re-used.
Redefine the settings storage area in TF-M as the non-secure storage
area, and put the storage partitions inside.

NCSDK-15340

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>