lib: nrf_modem_lib_trace: allow re-initialization
Allow re-initialization, let the driver return an error
if appropriate. The way the unit tests are written,
they expected the library to initialize the drivers
every time on initialization (which was not happening).

The tests should be updated eventually..

Additionally add a _deinit() function to nrf_modem_lib_trace,
since we need to be able to deinitialize the backends.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>