samples: lwm2m: v1.1 w/a TLV enabled bootstrap
With the LwM2M v1.1 the preferred content format would be SenML CBOR but
the current implementation doesn't tell it to a bootstrap server. In
those cases bootstrap uses TLV so there is need to enable it separately.
Usage of TLV is discouraged with v1.1. This w/a can be removed once
sending preferred content format functionality is implemented.

Signed-off-by: Veijo Pesonen <veijo.pesonen@nordicsemi.no>