applications: machine_learning: Add ext memory initialization module
This commit adds to the application module that configures QSPI and
external memory to be able to execute Edge Impulse library code from
XiP (Execute in place) memory. This is a part of application that adds
posibility to measure performance of the code execution from external
memory. That functionality is restricted only to nRF5340dk target.

To enable this functionality add `CONFIG_APP_ML_EXT_MEM` to your
Kconfig file.

Jira: NRFX-1829

Signed-off-by: Adam Wojasinski <adam.wojasinski@nordicsemi.no>