scripts/tools-versions-*.txt: Update to current versions
The PR updates the tools-versions{linux,win10,darwin}.txt files to use
the versions of the tools that we've been testing the toolchain bundles
with.

* For all three OSs, we removed gnuarmemb, SES, and ruby, and added
zephyr-sdk and nanopb.
* For Windows and macOS, we removed the GN version since in practice the
bundler will ignore it and retrieve the latest release.
* For macOS we changed the versions to those that will actually work
with the current bundler code. Since the bundler uses homebrew to
retrieve many of the tools, versions such as Python 3.8.2 were not
available and had to be updated. We also removed the version 'hashes',
since they were not used by the bundler code.
* For Linux, while the bundler currently ignores the file, we still
updated the versions to show what the bundler will retrieve.

Signed-off-by: Martin Galvan <martin.galvan@nordicsemi.no>