lib: nrf_modem_os: Use IRQ_PRIO_LOWEST instead of hard-coded 6
Use the lowest priority available in a given Zephyr configuration
instead of a hard-coded value of 6 (defined as `TRACE_IRQ_PRIORITY`
in `nrf_modem_os.c` and as `NRF_MODEM_APPLICATION_IRQ_PRIORITY` in
nrfxlib), as the latter is invalid for `CONFIG_ZERO_LATENCY_IRQS=y`.

For consistency with the IRQ used for modem traces, use also a local
definition of the IRQ number for application communication instead
of using `NRF_MODEM_APPLICATION_IRQ` from nrfxlib.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>