flash_patch: Add support for nRF52833
The nRF52833, like the nRF52840, supports disabling flash patch via
UICR.

Add it as a dependency for the Kconfig, and print the Cmake warning
when using nRF52833 with the bootloader. Add a new kconfig option
FLASH_PATCH_WARN to manage the logic of when to print the CMake warning.

Add the nRF52833 DK to the flash patch test.

Also specify in the Kconfig helptext that the relevant feature is
called the Flash Patch and Breakpoint (FPB) unit.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>