applications: asset_tracker_v2: Centralize check of nrf_modem_init ret
Centralize check of the `nrf_modem_lib_get_init_ret` return value.

Modem FOTA requires that the system is rebooted after a new modem image
has been applied indicated by the `MODEM_DFU_RESULT_OK`
return value from `nrf_modem_lib_get_init_ret`.

The application modules should not be allowed to initialize if
the modem is in this state where it expects a reboot to occur.
AT commands sent to the modem in this state fails, which causes the
various module `setup` functions to fail and trigger an assert.

The `cloud_module` will for instance try to retrieve the IMEI from
the modem during its call  to `setup`.

This patch also moves updating of the FOTA counter required for
LwM2M builds. This was failing due to the settings API not being
initialized when called from a `nrf_modem_init` callback handler.

Initializing the settings API in a `nrf_modem_init` callback handler
fails due to the settings API not finding the configured flash backend
which indicates that the flash backend is ready after `nrf_modem_init`
is called, which is set at init priority `POST_KERNEL`.

Fixes CIA-605

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>