tfm: Add option for allowing non-secure to reset the system
Add option for allowing non-secure to reset the system.
TF-M platform services cannot be called from interrupt context since
they are guarded by a mutex.
This is a problem when an fatal error has occured in zephyr and
the system has to be reset from the fault handler.

As a workaround add a configuration for allowing the non-secure to
reset the system and enable it by default.

NCSDK-15492

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>