samples: bluetooth: throughput: Improve user interface
Currently, role selection must be done with the buttons on
the board and the test is started on a terminal. It would be
nice to do everything on the terminal, e.g. when physical access
to the board is limited. Additionally ASCII art that shows progress
is updated to the current Nordic Semiconductor logo.

Ref: NCSDK-15579

Signed-off-by: Dominik Kilian <Dominik.Kilian@nordicsemi.no>