doc: gsg: Document environment variables for the Zephyr SDK
Since many of our users are updating from a previous version of NCS,
which defaulted to using GNU Arm Embedded as a toolchain, they have the
ZEPHYR_TOOLCHAIN_VARIANT set, and so they need to override it. To
simplify, just refer to the env vars themselves without going into
detail on whether they are really required.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>