applications: asset_tracker_v2: Handle modem functional mode requests
Handle modem functional mode requests from the security object:

During provisioning of a Pre Shared Key used in the DTLS connection to
either the bootstrap or the management server, the security object
requests via a callback that the modem must be put in a certain
functional mode.

To provision the PSK, the modem must be put in offline mode, and after
provisioning has been carried out, the security object requests
the modem to be put into normal mode to resume execution.

Implementing this callback gives the application control of when
this functional mode takes place. If not implemented, the security
object will put the modem into the respective state itself.

Fixes CIA-643

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>