cmake: Escape literal '"' chars in image_preload variables
The Zephyr build system supports passing CONFIG_ variables on the
west command line, like this:

    west build -- -D"CONFIG_SOME_STRING=\"some string\""

The logic in zephyr/cmake/modules/kconfig.cmake which implements this
feature takes each CONFIG_ variable defined in the cmake cache, and
writes the value of that variable directly into extra_kconfig_options.conf.
Therefore, if the user wants to add a *string* value to kconfig using this
mechanism, the extra_kconfig_options.conf entry would need to look like:

    CONFIG_BOOT_SIGNATURE_KEY_FILE="/foo/bar/private.pem"

and the image_preload.cmake definition that generates this result would
therefore be:

    set(CONFIG_BOOT_SIGNATURE_KEY_FILE "\"/foo/bar/private.pem\"" CACHE
        INTERNAL "NCS child image controlled")

However, as of NCS v2.0.0, this doesn't work correctly.  If I build with
this command line:

    west build --pristine --cmake-only -b nrf52840dk_nrf52840 -- \
        -D"mcuboot_CONFIG_BOOT_SIGNATURE_KEY_FILE=\"/foo/bar/private.pem\""

I see that the multi_image.cmake script from NCS generates invalid syntax
in both image_preload files:

    set(CONFIG_BOOT_SIGNATURE_KEY_FILE ""/foo/bar/private.pem"" CACHE
        INTERNAL "NCS child image controlled")

To fix this, NCS must add explicit escaping of '"' characters if they
are present in the user-supplied string.

Signed-off-by: Michael Scofield <mscof4378@gmail.com>