applications: asset_tracker_v2: add inactivity sample
This patch adds an additional sample after the sensor reports
inactivity.The use-case would be to set up a rather large movement
resolution (e.g. 1h) and an inactivity timeout to determine how long
we need to hold still for the status to be considered inactive (parking)
(e.g. 5min).
There can only be one additional sample for each step
in movement resolution, so we would expect the number of samples to
double.

Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>