lib: pdn: multiple callbacks for default PDP context
The current implementation only supported setting one
callback for the default PDP context, but multiple parts
of the application may want to receive these events.

Let the application register multiple callbacks for
the default PDP context.

Maintain a list of PDP contexts instead of allocating
them statically. Watchers of the default PDP context
are allocated an entry in said list.

PDN_CONTEXTS_MAX is removed, since now the number is
dynamic and depends on the heap.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>