tests: lib: at_sms_cert: remove wait for network registration
Do not wait for network registration as most of the at_sms_cert library
is application callbacks and do not interfere with the modem. Therefore,
there is no need to wait for the modem to be registered to the network
to test most of the library. Exception is the CMSS command that we
for now allow the modem to return an error.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>