test: lib: at_sms_cert: Add missing requirement for the test
The test from tests/lib/at_sms_cert/testcase.yaml requires a board
with a sim card on. However, it was not reflected in the test yaml
and the test was pick up by our hw setups and failing. This commit
fixes the test description by adding a fixture dependency.

fixes: https://projecttools.nordicsemi.no/jira/browse/NCSDK-15314

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>