matter: optimize QSPI NOR flash usage during DFU over SMP
Matter samples currently enable QSPI NOR flash on each BLE
connect because we thought there was not other useful
callback to enable it only for the SMP commands.

That resulted in needless flash power on when the BLE
connection was established for Matter commissioning.

Use mgmt_register_evt_cb() API, instead.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>