drivers: wifi: Fix low power mode checks
For normal mode the RD/WR status register access is mainly for
triggering/waiting, so, all return values are hard coded to zero, but
for low power mode the code relies on the return value read from the
RDSR value, so, this causes a false negative wakeup failure.

Fix the return values for public APIs to return the actual register
being accessed.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>