manifest: Update zephyr and TF-M to allow FPU with hard ABI
Update zephyr and TF-M with changes required to support FPU with hard
ABI.
- Zephyr: Allow enabling FPU and support not compiling NS application.
- TF-M: Support not compiling NS application.

Include the out-of-tree source files for the platform partition services
to be built by the zephyr build system instead of linking against
the platform_ns library.

Remove the dummy_uart_stdout workaround needed to exclude UART from the
non-secure application. This would provide a linker error in the
non-secure application when UART was disabled.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>