net: l2: wifi: Add nRF Wi-Fi management
This is based on Zephyr's Wi-Fi management but adds support for
wpa_supplicant, and eventually will be merged in to Zephyr's Wi-Fi
management.

wifi_shell.c is copy of Zephyr's, the original file is not used as
selecting that file automatically uses Zephyr's wifi_mgmt.

Bump the revision of sdk-hostapd to include corresponding changes.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>