doc: mcuboot: Clean up terminology for the 'direct-xip' mode.
The exact term 'direct-xip' should be used for the bootloader mode
where the application boots from the update partition directly
instead of swapping the images.
Remove 'XIP' abbreviation which may be confused with the
'execute-in-place' code execution mode from an external memory.
Additionally, insert link to the section which explains this mode.

Signed-off-by: Michał Grochala <michal.grochala@nordicsemi.no>