asset_tracker_v2: Workaround for multiple non-secure partitions
Partition manager has a bug that calculates the wrong size for the
non-secure storage span when it contains multiple partitions. It
it adds empty partitions in between the partitions so each becomes
the size of the SPU region size. However the empty partitions are not
taken into account, which triggers a Secure Fault in TF-M when trying
to access the other non-secure partition. This workaround increases
the size of the lwm2m_carrier partition so there is no need for an
empty partition in between the partitions.

Add the config to the lwm2m_carrier sample, so it's easy for customers
to see the config in case they need it.

Add new static partition for thingy:91 for lwm2m_carrier if
CONFIG_LWM2M_CARRIER is enabled.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>