tests: unity: support EXCLUDE argument to cmock functions
Introduce EXCLUDE argument for cmock CMake functions.
This allows tests to disable mock generation for functions that should
not be mocked or functions where hand-written mocks should be used.

The reason to skip generation of a mock is because CMock does
not have good enough support for variadic functions.
Related issue: ThrowTheSwitch/CMock#397

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>
Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>