applications: asset_tracker_v2: Implement nRF modem lib fault handler
Implement application-specific nRF modem lib fault handler to ensure
that the system reboots in case the modem faults.

Future improvement is to reinitialize the modem in case this
happens. But this is a larger effort because it ideally should be
aligned with reinitialization of the modem after modem FOTA.
Where the application currently also unnecessarily reboots.

Fixes CIA-578

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>