doc: nrf5340_audio: Add list of features
Add a table listing the features of the nRF5340 audio application,
with description, limitations and maturity for each feature.

The purpose of this is to give users a better overview of which
features exist, and how mature they are, in order to set expectations
correctly.

The table itself is put in the Software maturity levels page.  It is
duplicated into the top of the app doc using an include tag, so that
it is easily found and seen by the users reading the app doc.

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>
Signed-off-by: Grzegorz Ferenc <Grzegorz.Ferenc@nordicsemi.no>