net: lib: fota_download: Return proper error if FOTA init fails
Add a new error code FOTA_DOWNLOAD_ERROR_CAUSE_INTERNAL to be
used when any of the initialization phase fails. This should
reflect that the error cause is not coming from the network
or from the image itself.

When dfu_target_img_type() cannot detect the image type
or when  images are too large, they are rejected with
FOTA_DOWNLOAD_ERROR_CAUSE_INVALID_UPDATE

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>