cmake: remove duplicated variables in child image pre-load file
child-image specific settings are fetched both a regular and cached
variables. This causes some variables, for example those passed through
`-D<var>=<val>` to be in present in both lists and thus passed to the
child image twice.

Correct this by combining the two lists into a single one, and then
remove duplicates.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>