samples: nrf9160: modem_shell: fix for stopping GNSS in periodic mode
In application controlled GNSS periodic mode GNSS should be stopped
as soon as it gets the first fix. However, when A-GPS or P-GPS data
download was ongoing, stopping could be delayed because the used
work queue was executing the download. This commit moves starting
and stopping of GNSS from the modem_shell common work queue to the
system work queue to prevent this kind of delay.

MOSH-343

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>