net: lib: nrf_cloud: properly handle sequential connect calls
The flag that indicates if the connection poll thread is active
is set when the thread wakes upon taking the semaphore.
If the application makes sequential calls to connect, the
poll thread may not yet be awake.
Set the flag after the semaphore is given so that the
connect function returns the proper error value if
called again.
IRIS-4894

Signed-off-by: Justin Morton <justin.morton@nordicsemi.no>