doc fix: Incorrect reference to cmd.exe in Linux and MacOS
The guidance for installing the toolchain incorrectly referred to
`cmd.exe` in the instructions for Linux and MacOS. This changes
those references to terminal window and the Windows equivalent
to command-line window to not exclude PowerShell as an option.

Also fixes the links in the same section to properly use `links.txt`.

Signed-off-by: Wille Backman <wille.backman@nordicsemi.no>