drivers: wifi: Increase TX aggregation
With increase in QSPI clock, RPU is now faster than the Zephyr's
networking stack, so, the buffered packets are depleted faster leading
to lower aggregation (max 2 compared to 4 at lower QSPI frequency).

Decrease the amount of available token to maximize queueing in the host
to aggregate more (proportional to performance).

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>