bluetooth: Solving deadlocks and removing a memcpy
Peek to see what the next event going to be fetched is.
If it not possible to get a net_buf buffer for the event,
schedule to retry in 10ms. And then let other threads run.

Also, this commit removes the need for an extra memcpy
that was needed before. And also removes the extra
hci buffer that memcpy went into.