tests: net: lib: azure_iot_hub: Fix segmentation fault in test
Fix an segmentation fault that occurs due to uninitialized variable.
The test would crash randomly. It always crashed when disabling all
compiler optimizations.
Fix was to make getddrinfo return a valid pointer as result. This valid
pointer need to point NULL to avoid the Unit Under Test from
dereferencing it further for further operations. Previously this
pointer was uninitialized and was pointing some random value in memory
which may or may not be zero.

The test now also builds and runs well using the default (Minimal) C
library on qemu. Hence the overlay file for qemu is not needed.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>