sysbuild: disable partition manager / child image when sysbuild is used
Sysbuild and parent / child image are mutual exclusive.

Thus it is not possible to use both at the same time.
In parent / child image, the partition manager is invoked in the parent
image, and also in each domain image.

With sysbuild, the partition manager is invoked in sysbuild.
Therefore when building images through sysbuild, then partition manager
must be disabled in the image build.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>