net: lib: lwm2m_client_utils: Location assist result code and retry
Add result code handling for location assistance object. Result code
from server tells if the location should be tried again or not.
Currently -1 tells that there is something that needs to be fixed
in the server side and therefore attempting to obtain the location
again isn't done. If result is -1, the device needs to be rebooted
after the issues are resolved.
1 tells that there is a temporary issue in the server and client will
try again using exponential backoff.

When using GNSS object, it will prevent sending multiple requests
at the same time and will tell -EAGAIN that there is a request ongoing
and call me again after a while.

Signed-off-by: Jarno Lämsä <jarno.lamsa@nordicsemi.no>