net: lib: nrf_cloud: Implement optional P-GPS storage in external flash
Rewrite accessors to predictions so that pointers directly to
predictions are not used, but obtained from conversion functions.
If the predictions are in internal flash, that can be a simple cast.
If predictions are in external flash, one prediction at a time should
be read to a RAM buffer, then accessed from that.

Also fix issue that 0 predictions were requested from cloud.

Update logging of flash_area to get name from the underlying device
(since the structure changed).

When nrf_cloud_pgps_finish_update() is called, always unlock.

Jira: NCSDK-14963

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>