subsys: bl_monotonic_counters refactoring
This refactors the bl_monotonic counter subsystem.
There are two main changes in the sybsystem, firstly
the functions are now expecting a counter description
as a parameter, this allows other users later to use
the same APIs for counters by introducing new
descriptions. Also the APIs are now returning an
error code if they fail so that it is easier to
distinguish the cases where the counters cannot be
found.

Ref: NCSDK-9045

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>