dfu: dfu_target: Reset a properly downloaded image on dfu_target_reset()
There is now three ways to end the DFU flashing process:
* Mark the download as aborted.
  This is done by calling dfu_target_done(false).
  Continuing is possible at later time.
* Mark the downloaded image succesfull by calling dfu_target_done(true)
* Cancels or rejects existing image by calling dfu_target_reset()

Previously the cancellation was not actually implemented and it
only marked the image download as aborted. So effectively
there was no way to prevent flashing of a image that had been
downloaded and scheduled to be updated. Now reset() will erase
the area, or at least first sector, which prevents the update
and allows new image to be downloaded.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>