mcuboot: Don't support validating multiple images with NV counters
Don't support validating multiple images with NV counters as there is
no known valid use-case for this and it adds some complexity.

Also, provision the exact amount of slots that are configured instead
of adding one slot for alignment reasons. Doing something else than
what the user has asked for is dangerous and usually treated as a bug.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>