samples: matter: Added extracing config file name from CONF_FILE
Currently in CMakeLists there is a check that assumes CONF_FILE
to be always equal to exact prj.conf file name. However in some
cases it can be also absolute path to file on disc and the check
will fail.

Added extracting the exact config file name from CONF_FILE
value and used it in the check.

Signed-off-by: Kamil Kasperczyk <kamil.kasperczyk@nordicsemi.no>