bootloader: Fix bad rebase
Refactor the bootloader public key list handling to always operate
on a cmake list type instead of a comma-separated string.
This makes it possible to add the public key corresponding to the
current sign key to an empty list.
This would otherwise generate an error for the trailing or leading
comma in the --public-key-files argument.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>