bootloader: Allow giving an empty list of public keys for key revocation
Change CONFIG_SB_PUBLIC_KEY_FILES to treat the empty string as no
additional public keys for key revocation and validation.
Treat the string "debug" as the way to generate the two debug keys used
for testing.
Change the default of this option so that debug keys are not generated
by default.

NCSDK-18180

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>