ncs: default to newlibc nano
The nano flavour of newlibc used to be the default before NCS v2.2.0.

This changed when the commit in
https://github.com/zephyrproject-rtos/zephyr/pull/51340
was merged downstream, increasing the FLASH size of all nrf9160
samples by about 40KB, due to their requirement on some
variadic C functions whose implementation is much larger
in vanilla newlibc.

This patch reverts the change from upstream and maintains the
newlibc nano flavour as the default for nrf9160 samples,
and all other samples that use newlibc.

Removed decimal point accuracy from cloud codec tests.
The test was introduced while newlibc did not use nano
(and was failing when reintroducing nano).

Signed-off-by: Håvard Vermeer <havard.vermeer@nordicsemi.no>
Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>