mpsl: fem: nrf21540_gpio_spi: fix built-in model init
The MPSL-based temperature driver is not reentrant. Current
implementation of the nRF21540 GPIO+SPI built-in model calls it from a
different execution context than other modules. That leads to preemption
of the temperature driver that effectively locks the entire application
in an endless loop. This commit unifies the execution context of the
temperature driver in the nRF21540 GPIO+SPI built-in model and fixes the
bug that results from this preemption.

Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>