net: lib: azure_iot_hub: Fix possible race in polling thread
In certain situations, there could be a race between a thread
disconnecting from for instance DPS and updating the state, and
the polling thread checking the state.
This could lead to the polling thread continuing to poll on
a dead socket, leaving it unresponsive for the duration of
the configured MQTT keep-alive.

This patch changes the state checks and introduces intermediate
transport-level state that distinguishes between connection
establishment on transport level (that does need polling) and
MQTT level (that does require polling).

Fixes CIA-846

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>