cmake: variant build now re-uses original configuration
Fixes: NCSDK-18108

Instead of passing on settings for variant builds then simply copy
existing .config and autoconf.h header for re-use.
This ensures that any changes done in the original image are
automatically re-used in the variant build.

Because files are now copied then we must amend
CONFIG_NCS_IS_VARIANT_IMAGE to autoconf.h to ensure it's picked up by
the adjusted linker script template.

Because CONFIG_NCS_IS_VARIANT_IMAGE is passed as CMake variable we
can remove its definition in Kconfig and thus avoid misuse of the
setting as well as user confusion.

Disable menuconfig and friends for variant builds.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>