applications: asset_tracker_v2: Remove fresh flag for dynamic modem data
Remove fresh flag for dynamic modem data. We can never guarantee that
data gathered in a sample request actually arrive at the broker.
We have retransmission functionality via the QoS library for MQTT based
clouds, but retransmission are limited to x amount of attempts.

The fresh flag is associated with individual parameters sampled
on the device that is not equal to the last sampled value. If the device
considers a sampled value as old, it will not be included in the next
cloud publication. The problem is that we don't know if the value that
was sampled the previous sample cycle was actually sent.

This functionality is application specific so the reasonable path is
to help customers (if there is a demand for it) to implement this kind
of functionality in their application. Rather than having it in
the Asset Tracker v2 as a standard feature.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>