modules: Mcuboot allow imgtool extra params
Pass the value in CONFIG_MCUBOOT_EXTRA_IMGTOOL_ARGS to imgtool.py.

This is necessary because we are maintaining a copy of the build
scripts for MCUBoot signing. Upstream already supports this option.

There are many more "imgtool.py sign" options than what we have
encoded in our build system, and this option makes it easier for users
to use these options without patching the build system.

Ref: NCSDK-9045

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>