bl_storage: Change the monotonic_counter API
The montonic_counter implementation has some code for supporting
multiple counters, but the API only supports one counter. This commit
updates the API and finalizes this support, but still does not add
another counter.

Introduce a counter description into the monotonic_counter API to be
able to distinguish between multiple counters.

Also introduce some error handling in case the counters are not found.

bl_validation uses the affected API and has been updated. It's API
also needed to be updated to deal with new error handling.

Ref: NCSDK-9045

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>