bl_storage: provision: Require an even number of counter slots
We have been adding an extra counter slot if the user specified an odd
number of counter slots. This results in the configuration not
matching the actual implementation.

It is bad practice to allow the user to ask for one thing, and give
them something else, so instead we require that the user specifies an
even number of slots.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>