cmake: multi image: place results from ncs_file() in cache
When using ncs_file() for looking up configuration and overlay files for
child images, then the result is returned in a local scoped variable.

To ensure correct behavior between CMake invocations then this result
must be stored in the CMake cache.
To further increase the robustness of the build system then the local
scoped variable is cleared so that only the CMake cache variable
remains.

This fixes an issue where Kconfig fragments or devicetree overlays for
a child image would only be applied on first CMake invocation but not
on subsequent invocations because the file(s) would not be stored in the
CMake cache and the algorithm for applying child image board specific
files are only executed on first run.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>