bl_storage: Omit counter functions when counters are disabled
Previously, counter functions like num_counter_slots and
get_montonic_counter were returning 0 when counters were disabled.

Running code for accessing a counter when counters are disabled
indicates a bug in the FW and there is no reason to support this.

We ifdef out the code in this patch.

Now a link failure will occur for such misconfigurations instead of
having functions that appear to work, but just return 0.

This also saves space for the case where CONFIG_SB_MONOTONIC_COUNTER
is disabled.

Ref: NCSDK-9045

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>