tests: Cleanup setUp and tearDown functions
The _Init() and _Verify() are not not needed to be placed in setUp() and
tearDown() functions in the unit tests. These are automically called by
the generated function CMock_Init() in runner_*_test.c.

The test_suiteTearDown() is not needed as there is a weak
definition of the same function doing the same thing in
generic_teardown.c within unity framework.

Used double quotes instead of angular brackets when including cmock
generated headers because
- They are not system headers or library headers to be included that way
- They are generated in subfolders of current working directory
- The _Init and _Verify functions are not automatically called by Cmock
if these headers are not included this way. I believe the
generate_runner.rb assumes that all cmock files are included with double
quotes and not angular braces which is a normal way of including these
headers.

Fixed a couple of test bugs in modem_jwt test and location_module test
that were uncovered due to these changes.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>