manifest: Update Matter sdk revision that includes factory data fix
In the newest version of the Matter SDK, there is a fix for
the factory data set that allows using the "user" data field
in code. There is also updated documentation for creating
a factory data partition.

Added a known issue for NCS 2.2.0, 2.1.2, and 2.1.1 that are
impacted by that issue.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>