Improvement inside ESB driver to reduce fly time between emission and reception. These modification achieve under 100µs fly time with 2MBPS transmission speed, 8bit CRC, no acknowledgement.
esb_config struct has been added 2 options :

- optimize_txen_delay
When this feature is enabled, radio TXEN delay is
cut off from 130µs to 40 µs.
The radio peripheral needs some time to start-up
analogs components of the radio. On nRF51 and nRF24L series,
a hard coded 130 µs delay is implemented. If ESB connection
is achieved only between nRF52 devices, this delay can be cut off
to 40 µs.

- never_disable_tx
When this feature is enabled,
if mode is NRF_ESB_MODE_PTX and if packet
is not acknowledged, then radio peripheral will
remain in TXIDLE state instead of TXDISABLE when
transmission is pending. This reduce delay between
consecutive transmissions but consumes more energy.