net: lib: nrf_cloud: Fix P-GPS issue with non-zero offset in ext flash
When an application configures P-GPS and another subsystem to both use
external flash, the P-GPS partition can often start at a non-zero
offset.

Fix a mistake in how flash_area_read() is used -- the offset must be
from the start of the partition, not the start of the flash device.

Jira: NCSDK-19097

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>