bl_storage: Make NSIB aware of whether it supports NV counters
Currently NSIB/B0n is not aware of whether or not it supports NV
counters.

This is unfortunate as it is then forced to include code for it even
when it doesn't support it.

Pass on a Kconfig from the parent image to NSIB/B0n to make it aware.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>