application: serial_lte_modem: Shorten UART recovery delay
When SLM detect UART DISABLED by breaks or framing errors, it
tries to recovery with a delay. However the delay of 500ms is
too long and could miss first data from MCU after MCU resets.
Update the delay to 10ms.

Also add a PULLUP for UART_2.RX pin when used.

Signed-off-by: Jun Qing Zou <jun.qing.zou@nordicsemi.no>