drivers: sensor: bh1749: Add wait-for-ready functionality
The BH1749 requires some initialization to happen iwith a certain
delay, and hence the driver achieves that using delayable work items
on the system workqueue.

In some cases this results in a race where an application may
attempt to use sensor APIs before the sensor is fully initialized.
This can cause issues both with the sensor and subsystems that
the driver uses, for instance GPIOTE.
In turn, this could cause the application to hang indefinitely
in an interrupt handler. See nRF9160 errata 4 for more details.

This patch introduces a configurable wait time for the
driver to complete initialization before any public API
is usable.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>