scripts: cred_store: Add cred store script
This commit introduces a new set of scripts for managing certificates
stored in the modem on a cellular device. It's mainly a wrapper around
the AT-command %CMNG with the exception of %KEYGEN. It also uses
pythons serial library to communicate with the modem. This makes it
possible to manage certificates in the modem from the command line in
a cross-platform manner.

Execute cli.py to use it from the command line:
python3 cli.py /dev/tty.usbmodem0009600834341 list

To get the full help message use --help:
python3 cli.py --help

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>