bluetooth: adv_prov: Expand advertising state
Added new state fields to the general API of the Advertising Provider
modules. They are used to notify the registered providers about the
new advertising session and the RPA rotation state.

Integrated the new API with the Fast Pair advertising provider. Added
a Kconfig option for dropping the Fast Pair discoverable payload on
RPA rotation.

Ref: NCSDK-15502

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>
Co-authored-by: Marek Pieta <Marek.Pieta@nordicsemi.no>