cmake: correcting path for protoc when using NCS Toolchain
Jira: NCSIDB-816

Move location of `env_list_delimeter` to allow it to be used also for
PATH handling in custom commands env.

This ensures that the CUSTOM_COMMAND_ENV includes local core tools in
path thereby ensure proper behaviour of the protoc when the
linux shell script wrapper is used as the protoc-gen-nanopb plugin.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>