bluetooth: controller: Fix unexpected low TX power with FEM connected
Previously the initial default output power would be selected as if
there was no FEM connected. For example, suppose that the application
requested 20dBm power at the antenna. Even though such power could be
achievable using the configured front-end module, that power would first
be converted to the maximum output power supported by the SoC,
which usually is much lower. Then, when the device transmitted frames,
this decreased power would be treated as the requested
power and divided between the radio and the external PA.

This commit fixes that by setting the default radio output power when
enabling Bluetooth, when the FEM driver is already initialized.
This way the controller takes FEM into account in its initial output
power calculation.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>