net: lib: lwm2m_client_utils: Add support for connection pre-evaluation
Connection pre-evaluation estimates the LTE radio signal state and
energy efficiency in a cell that the modem would use for data
transmission. Application can use the information to decide when
to start the data transmission.
If estimated energy usage is too high the application pauses the
lwm2m client and waits for better signal state.
User can set evaluation period in milliseconds and maximum time
in seconds to pause the transmission.

Signed-off-by: Juha Ylinen <juha.ylinen@nordicsemi.no>