west.yml: fix hostap project name
The preferred name of a project in our west.yml file is the name of
the repository as it is commonly referred to, or its usual upstream
git repository name, and not the name of the NCS fork. The hostap
project doesn't abide by this convention, so clean it up.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>