west.yml: add userdata for tracking OSS upstreams
Not all OSS repositories that are present in both NCS and included in
zephyr/west.yml are forked from the zephyrproject-rtos GitHub
organization's fork of the relevant repository.

Add information on repositories which are forked from other locations
using west's 'userdata' field. This will allow our internal tooling
for automated tracking of out of tree patches to improve its accuracy.

This feature was added in west v0.13.0, so bump the version field in
the manifest accordingly.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>