manifest: Check out qcbor version for TF-M and enable attestation
Add an entry of the TF-M dependency QCBOR and supply it to the
TF-M build system.
In zephyr this dependency is no longer included due to zephyr project
not having resolved if the license of this project is OK to use.
Re-enable TF-M initial attestation partition which is dependent
on the QCBOR functionality.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>