mpsl: cx: MPSL_CX is disabled by default
In the current implementation, CONFIG_MPSL_CX is enabled whenever WiFi
and either Bluetooth LE or 802.15.4 are enabled in the same application.
However, this approach is misleading in case of nRF5340. The nRF5340
network core can never satisfy this condition as it only handles short
range and cannot have WiFi enabled. As a result, CONFIG_MPSL_CX must be
enabled explicitly by the user for the network core. Without it, it's
easy to end up with an invalid configuration where CONFIG_MPSL_CX is
enabled automatically for the nRF5340 application core image and it is
disabled for the nRF5340 network core image.

This commit changes the default so that unless explicitly enabled,
CONFIG_MPSL_CX is always disabled. Configuration needed for both cores
to enable coexistence becomes consistent, at the cost of additional
options to be selected manually for the application core.

Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>