west.yml: split contents into submanifests
We are constantly running into merge conflicts when two projects that
are close to each other in west.yml change revisions, as git
identifies changes to e.g. the two revision fields as conflicts in the
same hunk.

To avoid this in the future, split up the contents of west.yml into:

- individual YML files for each NCS repository
- another YML file for the third-party repositories we include in NCS

This also lets us clean up a few issues in west.yml, namely:

- some NCS repositories are incorrectly listed in the "third-party"
  section of west.yml (cirrus, openthread): split into individual
  files

- remove some now-unused remotes

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>