scripts: west_commands: add helpers to extract userdata
These will make it convenient to use the userdata in our manifest.

I wanted to use a dataclass for NcsUserData, but I'm hitting a null
pointer exception where the internals of the dataclass module are
trying to dereference None. This looks like a bug in CPython. I don't
have time to figure this out, so a NamedTuple will do for now.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>