scripts: west ncs-loot: work with all repositories
Rework the command implementation to use the manifest userdata fields
previously added to west.yml. This allows us to print the out of tree
patches for all sorts of repositories with known upstreams, not just
the ones we inherit from Zephyr.

This is a good opportunity to clean up the code and improve the
human-readable output as well.

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>