manifest: Update Matter revision to pull CHIPtool workflow fix
Currently CHIPtool (Matter controller for Linux) has dependency
to openSSL.
The issue is that Ubuntu 20.04 and Ubuntu 22.04 use different
versions of OpenSSL and the version built under 20.04 does not
work under 22.04 (without some hacks to install the old openssl
version) and vice verse.

Removed the dependency to the dynamic library to build CHIPtool
with the mbedTLS sources.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>