doc: lib: lwm2m_carrier: fix custom initialization documentation
The documentation stated that the lwm2m_carrier_custom_init function
is to be called from within the event handler, which is incorrect, as
it is to be called in the library thread prior to lwm2m_carrier_run.
Added missing space in the lwm2m_carrier sample description.

Reference: NCSDK-19890

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>