samples: wifi: shell: Refactor memories to get more kernel HEAP
When running Zperf its observed that nrf700x driver struggles to get
memory for crucial messages exchange with RPU. This used to work
reliably with 200000 but due to recent changes the HEAP consumption has
increased.

So, decrease the TX buffers as they are the costliest and consume
2 buffers/packet, to make room for more HEAP.

Once we do HEAP usage analysis these numbers can optimized.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>