drivers: wifi: Fix the memory leak
This overwriting of status is causing a memory leak, earlier to the
driver clean up commit 3384e56be5 ("drivers: wifi: Fix memory leak in
negative cases the buffers") buffers were freed immediately after
failure, so, this wasn't affecting, but now that we have moved the
buffer free to the the FMAC entry point, this overwrite of status causes
memory leaks.

This hack was introduced to handle an issue wit Linux networking stack,
so, not needed for Zephyr in the first place.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>