manifest: Update Zephyr revision
Pull in changes in Zephyr that fix a problem in the system timer
driver (nrf_rtc_timer) that in specific conditions the COMPARE
event for a scheduled timeout was not handled properly and in
consequence the timeout expired with an additional delay of 512
seconds, i.e. after the RTC overflowed.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>