tests: bluetooth: Make memcpy operation more clear
Previously the memcpy was copying the contents of the address of
property.current assuming that 'current' will always be the first
element in the structure. This is now fixed by changing to source
of the memcpy to address of the property struct itself. Technically they
are both the same, but this makes the intention more clear and makes the
code less prone to bugs.

Found as a voilation of MITRE, CWE-131 by sonarcloud.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>