drivers: wifi: Fix data access violation
When running Zperf UDP-RX and issuing "wifi status" in the background
few times causes a crash (data access violation), the callstack is not
helping.

The current theory is that, during "wifi status" the driver waits for
RPU responses which are in the same queue as RX frames, so, this
causes delay in RX frames processing and network stack runs out of
buffers and subsequent allocations in the driver wait for memory to be
reclaimed, and this internally uses z_pend_curr that causes to switch
the thread and crash is seen in the WPA supplicant thread context
(probably in l2_packet_recv or socket select).

There is a similar issue reported in [1] but the fix didn't help, so, as
a temporary workaround disable the timeout for allocation, with this the
issue is not seen.

[1] - https://github.com/zephyrproject-rtos/zephyr/issues/29003

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>