bluetooth: hids: Change HID report map size from uint8_t to uint16_t
In the bluetooth HID service, the size of a report descriptor is held
by a uint8_t, which means that any report descriptors can have a length
of 255 bytes at most. Most practical applications of report descriptors
demand a more complex structure that far exceed that length. This change
increases that limit to an uint16_t.

Signed-off-by: Jacob Sánchez <jacobszpz@protonmail.com>