applications: asset_tracker_v2: Disable logger for printk()
Disable the use of the logging sybsystem for printk().
This is done as a workaround for an issue where long strings may
not be printed properly.
The downside of this solution, is that prints done using printk()
may be out of sync with those printed with the logger in
deferred mode.

Fixes CIA-892

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>