lib: nrf_modem: do not reset poll signals
Don't reset poll signals on `ZFD_IOCTL_POLL_UPDATE`, it is not
what we are supposed to do there. Just report the flags instead.

Simplify the poll callback logic, and simply raise the flags
as reported by `revents`. Those are set only if they were set
in `events` to begin with, or they are `POLLHUP`/`POLLERR`/`POLLNVAL`
in which case we must report them unconditionally.

The ZSOCK_POLL values are build-asserted against NRF_POLL values.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>