lib: nrf_modem: deprecate NRF_MODEM_LIB_IPC_IRQ_PRIO_OVERRIDE
This Kconfig was meant to let us easily override the IPC IRQ
priority NCS-wide, in case that was needed, to smooth the transition
from using the EGU interrupt to using the IPC interrupt only.

That proved to work okay, and this option is no longer needed.
The application can override it via DTS for application-specific
use cases.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>