mpsl: cx: allow to use nRF700x with nRF53 without nRF RPC
Some applications might not want to have the possibility to enable or
disable coexistence in run-time. Currently, serialization of these APIs
is enforced and cannot be turned off when nRF700x coexistence interface
is selected. This commit changes that by allowing the user to disable
this serialization. It is still enabled by default though.

Signed-off-by: Jędrzej Ciupis <jedrzej.ciupis@nordicsemi.no>