drivers: wifi: Fix TWT flows dump
The TWT flow info structure between Driver and RPU and Driver and Zephyr
are not same, so, cannot use memcpy, instead need to copy individual
fields. As we now need to use converion to and from ms and float, add
helpers functions to simplify the process.

Fixes TWT flows part of "wifi ps" command output.

This also fixes the interval calculation in the TWT event.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>