applications: asset_tracker_v2: move secondary partition to ext flash
Add necessary config to use external flash for mcuboot_secondary. The
external flash was not available on the nRF9160 DK until v0.14.0. This
will free up a significant amount of internal flash.

MCUboot also needs the SPI_NOR flash driver and config to be able to
access the external flash to perform the swap between external and
internal flash.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>