drivers: wifi: Fix free after use under load
Under load if semaphore for waiting on RPU event is timedout then the
function for signal_poll returns but the RPU events in queue are still
executed later which try to work on signal_info which is in WPA
supplicant stack but no longer valid (function returned), this causes a
data exception.

Fix this by checking the semaphore return value for timeouts and by
explicitly making the signal_info NULL before return and checking for
that before access.

Fixes SHEL-744 instead of working around.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
Signed-off-by: Krishna T <krishna.t@nordicsemi.no>