lib: dk_buttons_and_leds: fix for nRF9160DK IO expander
The driver for the PCAL6408A IO expander on the nRF9160DK does not
support disabling the interrupts while they are being processed.
Move disabling of the interrupts to the buttons_scan_fn function.
Also, the IO expander driver only supports edge driven interrupts, so
this is added to the configuration.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>