samples: nrf5340: Remove empty net core from multi-core sample
Due to a change for all nRF5340 samples from 82a3b99 an empty net core
image was added to the multi-core sample. In this sample we provide our
own net core image and don't need the empty net core image for the
build.

Added override to the KConfig and a configuration to the sample so that
similar issues may be caught by CI.

Ref. NCSDK-20186

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>