partition_manager: fix static partition in start-to-end region
Fix bug in partition manager where it was not allowed to have a packed
static partition in the start of a START_TO_END region unless it filled
the entire region. This checked if the first gap between static
partitions started at the end of the region. Resulting in static
partitions have to fill the entire region to pass the verification.

This fix changes the check to compare the end of the first gap against
the end of the region. Which allows non-static partitions to be placed
after the static partitions.

Fixes NCSDK-20630

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>