manifest: nrfxlib with import key + TF-M minor fix
This makes sure that we always call the function
psa_import_key_into_slot when no crypto accelerator
import_key functionality is available. This is needed
because our Oberon library does not support import_key
for symmetric keys but it supports doing operations
with symmetric keys. The psa_import_key_into_slot
function always includes the logic for symmetric keys,
which has minor effects in increasing the code size.
The support for asymmetric keys on the other hand is
configurable which is a good for us since Oberon does
support import_key with asymmetric keys.

It also brings a build warning fix for
sdk-trusted-firmware-m when the encrypted
ITS is not enabled.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>