drivers: wifi: Increase priority for IRQ and BH workqueues
Meta-IRQs are the highest priorites that is designed to be used by
botttom halves, so, use it for IRQ WQ (though we call IRQ
this is not the original interrupt handler, but called from QSPI
interrupt handler). Using this or any other co-operative priority for
BH WQ causes issues with RX (less scan results), so, we leave it at
highes preemptible priority.

With this we need to increase the workqueue items and also the network
management events to avoid starvation as now the RX runs on high
priority.

Signed-off-by: Krishna T <krishna.t@nordicsemi.no>