samples: nrf9160: https_client: Wait for PDN to fully activate
With the default (IPV4V6) dual stack packet data network (PDN), the
network is not fully up until you have allowed a chance for the
PDN_EVENT_IPV6_UP event to occur. Wait for the event before looking up
the address, with a time out in case the carrier only supports IPv4.
If you do not wait then there is a race condition, and if you do a
lookup immediately on PDN_EVENT_ACTIVATED you may get the IPv4 address,
even though the network supports IPv6 (on all subsequent lookups). If
the PDN defaults are overridden to IPV4 only, then skip the wait, as it
will always time out.

Signed-off-by: Sly Gryphon <sgryphon@computer.org>