Bluetooth: controller: Allow CONFIG_BT_MAX_CONN=1
BT_MAX_CONN defines the number of host contexts. It seemed like a good
idea at the time to piggyback on it for setting the number controller
central/peripheral contexts, but that is not correct:

There is a valid use-case for an application that wants to be central or
peripheral, but not simultaneously. In that case, enforcing this 1-1
relationship would waste memory in the host.

The upstream controller has no such restriction, so that also clears up
build system headaches.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>