applications: asset_tracker_v2: Use correct device ID length
Use correct device ID length when populating the Azure IoT Hub
configuration structure. The length of the client ID should not include
the null character. This is required so that Azure SDK is able to
properly construct the MQTT username. Both length and pointer to client
ID buffer is now explicitly set after HW ID has been obtained.

Set client ID length buffer to `HW_ID_LEN` macro that will change
depending on the type of HW that is used. This was previously hardcoded
to 15 (IMEI length) and would not work for UUID HW IDs.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>