samples: tfm: provisioning images rewrite
This splits the provisioning image to two
different samples. One for the application core
and one for the network core for the NRF5340.
This is done to comform with how the other NCS
samples work. It is easier for customers to use a
separate sample as a child image instead including
the child image as a Zephyr module.

Ref: NCSDK-18228

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>