doc: add "ncs-tool-version" custom role
Added a custom role, `ncs-tool-version`, that collects tool
version information from pip, .txt and .yml files and
displays a given tool version.

This role is meant to replace the tool display mechanism found
in doc/nrf/versions.txt.in, with the argument for a given tool
mirroring the CMake replacement symbol.

Example of use:
:ncs-tool-version:`CMAKE_VERSION_WIN10`

Signed-off-by: Gaute Svanes Lunde <gaute.lunde@nordicsemi.no>