bluetooth: fast_pair: Fast pair advertising requires entropy generator
Fix missing select of entropy generator, which is used by fast pair
advertising service and fast pair gatt service.
This works in the default config which uses NRF_SECURITY that has this
select, but with the Oberon backend (CONFIG_BT_FAST_PAIR_CRYPTO_OBERON)
enabled this fails with a compilatior error for missing sys_csrand_get.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>