samples: use a separate random per peer
If multiple devices initiate a ranging to the same peer at the same
time, this could result in the attempts interfering with each other.

This can for example happen if an advertiser is scanned by multiple
devices at the same time. In that case, all scanners receive the scan
response from the advertiser and start a ranging at the same time.

Added a per-peer random based on its address, such that this should
not happen. The random is passed as input to the ranging library.
This makes sure that only the pair of devices that have the same
random can range with each other. This should limit the interference
from other devices trying to range to the same device at the same time.

Signed-off-by: Jan Müller <jan.mueller@nordicsemi.no>