tfm: Default off brainpool curve types not supported by PSA drivers
Default off brainpool curve types not supported by PSA drivers.
These do not have mbedtls builtin driver types, and instead rely
on MBEDTLS ECP legacy implementation.

The default for these are off when not building for TF-M so it makes no
sense to include them here.

Default these to off as we don't have support in CC3XX or Oberon drivers
for these curve types.

They are also listed as not supported in driver_config.rst.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>