asset_tracker_v2: tests: Make json_common use unity instead of ztest
The json_common unit test is now made to use the unity test framework
instead of ztest. This is done to align it with the other unit tests
in the asset_tracker_v2 folder that also use unity instead of ztest.

Replaced all zassert functions with the corresponding TEST_ASSERT
macros in unity.

The message parameter of the assert functions were removed as they
were superfluous. If the test fails, it will indicate which assert
failed.

Fixed a segmentation fault that occurred because we were
invoking cJSON_Delete() on random pointers.  This is now fixed by
memsetting the dummy variable to 0 in tearDown(), which clears all
pointers. The cJSON_Delete() will not try to free the pointer if it
is NULL.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>