driver: wifi: Enforce minimum value of NRF700X_MAX_TX_TOKENS
The minimum value of NRF700X_MAX_TX_TOKENS has to be atleast
WIFI_NRF_FMAC_AC_MAX (5). If not, this will lead to the setting of
num_tx_tokens_per_ac to 0 by wifi_nrf_fmac_init() which in-turn will
lead to divide by zero errors in clear_spare_desc_q_map() and
set_spare_desc_q_map(). This is now fixed by enforcing a minimal value
for the KConfig symbol NRF700X_MAX_TX_TOKENS.
The max value is set to 12 as we will not support more than that.
Also added ASSERT statements to catch this condition if it ever arises.

Found as violation of CERT, INT33-C and MITRE, CWE-369 by sonarcloud.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>