applications: serial_lte_modem: slm_at_host coverity
Possible non null-termination in the final result.
This was reported by coverity.
Refactored the function, replaced memcpy for the CRLF
with assignments as those were confusing the coverity and
should be cleaner. Replaced strstr with own implementation of
strrstr to find the last occurrence of the result insted of
finding all the occurrences, also added a check for the buffer
length so that it doesn't overflow with a very long response when
inserting CRLF in the end.

Signed-off-by: Jarno Lämsä <jarno.lamsa@nordicsemi.no>