bootloader: cmake: keys: Change from source dir to app config dir
Changes the base directory for signing keys to be based on the
application configuration directory instead of the source directory
(this has no change on projects not setting the application
configuration directory). This enhances security as signing keys
should not be the same for multiple boards, each should be unique.

Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>