applications: asset_tracker_v2: Fix race for credential provisioning
The Memfault SDK provisions credentials on devices. For nRF91,
this needs to happen after modem initialization and before
functional mode is set to normal.

Previously there was a race because both the Memfault integration
code and the modem module used NRF_MODEM_LIB_ON_INIT hooks
to trigger their actions.

Now this is changed so that the modem moule does both operations
sequentially. It's not optimal to let the modem module handle
the Memfault credentials this way, but the alternatives caused
a lot of added code and events between the modem module and debug
module, which did not seem worth the trade-off.

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>