manifest: Bugfix for external flash with single slot MCUBoot
This is a bugfix where you can't update the network core using only a
single secondary slot in external flash. This is due to the fact that
the network core update data is inaccessible to the network core in
external flash and needs to be copied into the ram of the application
core.

Ref: NCSDK-21379

Signed-off-by: Sigvart Hovland <sigvart.hovland@nordicsemi.no>